group duplicate toasts
