--- docs/chunk-EFJEUBLF.js
@@ -1,66 +1,0 @@
-import{$a as C,Aa as L,Cb as Ge,Db as He,Ea as E,Eb as Re,Ja as x,Ka as ue,La as c,M as T,Ma as de,N as B,Na as w,O as ae,Oa as V,Pa as Me,Q as D,Qa as Ae,Ra as s,S as xe,Sa as r,Ta as m,Ua as U,Va as g,Wa as v,X as Z,Xa as Fe,Y as M,Ya as Ie,Z as le,Za as Ne,_,_a as a,a as y,ab as ce,b as S,bb as A,cb as F,da as J,db as I,ea as j,eb as N,fa as G,fb as O,ga as we,gb as Oe,ha as Te,hb as ke,i as X,ia as H,jb as Pe,k as Se,ma as R,na as Q,ob as Be,p as Ee,qb as je,sb as he,ta as De,v as Ve,wa as d,xa as p}from"./chunk-RBV3YYQO.js";var Le=(()=>{class t{constructor(){this.oscillatorType="sine",this.frequency=430,this.audioCtx=new AudioContext}notify(){this.audioCtx.state==="suspended"&&this.audioCtx.resume();let e=this.audioCtx.createOscillator(),n=this.audioCtx.createGain();e.type=this.oscillatorType,e.frequency.value=this.frequency,n.gain.setValueAtTime(.3,this.audioCtx.currentTime),n.gain.exponentialRampToValueAtTime(9e-5,this.audioCtx.currentTime+.4),e.connect(n),n.connect(this.audioCtx.destination),e.start(0),e.stop(this.audioCtx.currentTime+1)}static{this.\u0275fac=function(n){return new(n||t)}}static{this.\u0275prov=B({token:t,factory:t.\u0275fac,providedIn:"root"})}}return t})();var me=class{constructor(i=1/0){this.capacity=i,this.storage=[]}setCapacity(i){this.capacity=i}getCapacity(){return this.capacity}enqueue(i){if(this.size()>=this.capacity)throw Error("Toasty queue has reached max capacity, you cannot add more items.");this.storage.push(i)}dequeue(){return this.storage.shift()}size(){return this.storage.length}getElements(){return this.storage}removeId(i){let e=this.storage.findIndex(n=>n.id===i);if(e!==-1){let[n]=this.storage.splice(e,1);return n}}},h=function(t){return t.Basic="basic",t.Success="success",t.Error="error",t.Info="info",t.Warning="warning",t.Custom="custom",t}(h||{}),k=(()=>{class t{constructor(e){this._toastySoundService=e,this.TOASTY_SERVICE_CONFIG={capacity:10,duration:5e3,messages:{now:"just now",seconds:"{0} seconds ago",minutes:"{0} minutes ago"},threshold_offsetX:170},this.queue=new me(this.TOASTY_SERVICE_CONFIG.capacity),this.counter=0,this.newToastBehaviorSubject=new X([]),this.newToast$=this.newToastBehaviorSubject.asObservable(),this.deleteToastBehaviorSubject=new X(void 0),this.deleteToast$=this.deleteToastBehaviorSubject.asObservable(),this.updateToastBehaviorSubject=new X(void 0),this.updateToast$=this.updateToastBehaviorSubject.asObservable()}setDefaultDuration(e){this.TOASTY_SERVICE_CONFIG.duration=e}getDefaultDuration(){return this.TOASTY_SERVICE_CONFIG.duration}setCapacity(e){this.TOASTY_SERVICE_CONFIG.capacity=e,this.queue.setCapacity(e)}getCapacity(){return this.queue.getCapacity()}showToast(e,n,o,l){this.counter++;let u=o?.sticky?1/0:o?.duration??this.TOASTY_SERVICE_CONFIG.duration,f={id:this.counter,title:e,message:n,type:o?.type??h.Basic,timestamp:Date.now(),duration:u,expires:l?1/0:Date.now()+u,config:o,get icon(){return o?.type==h.Success?"\u2705":o?.type==h.Info?"\u2139\uFE0F":o?.type==h.Warning?"\u26A0\uFE0F":o?.type==h.Error?"\u274C":""}};return this.queue.enqueue(f),this.newToastBehaviorSubject.next([...this.queue.getElements()]),f.config?.beep&&this._toastySoundService.notify(),f.config?.sticky||(f.config?.progressBar&&setTimeout(()=>{this.updateToast(f.id,{animatePbar:!0})},100),l||setTimeout(()=>{this.removeToast(f.id)},u+5)),this.counter}showToastPromise(e,n,o){let l=this.showToast(e,n.loading,S(y({},o),{loading:!0}),!0),u=o?.duration??this.TOASTY_SERVICE_CONFIG.duration;n.promise.then(f=>{let b=typeof n.success=="function"?n.success(f):n.success;this.updateToast(l,{message:b,type:h.Success,expires:Date.now()+u,config:{loading:!1}})}).catch(f=>{let b=typeof n.error=="function"?n.error(f):n.error;this.updateToast(l,{message:b,type:h.Error,expires:Date.now()+u,config:{loading:!1}})}).finally(()=>{o?.sticky||setTimeout(()=>this.removeToast(l),u)})}removeToast(e){this.newToastBehaviorSubject.next([...this.queue.getElements()]),setTimeout(()=>{let n=this.queue.removeId(e);this.newToastBehaviorSubject.next([...this.queue.getElements()])},500)}closeToast(e){let n=this.queue.removeId(e);this.newToastBehaviorSubject.next([...this.queue.getElements()])}updateTouchStart(e,n){let o={date:Date.now(),touchStart:n};this.updateToast(e,{gesture:o})}updateTouchEnd(e){let n=this.queue.getElements().forEach(o=>{if(o.gesture){let l=Math.abs(e.clientY-o.gesture.touchStart.clientY),u=Math.abs(e.clientX-o.gesture.touchStart.clientX);Date.now()-o.gesture.date<500&&u>this.TOASTY_SERVICE_CONFIG.threshold_offsetX?this.closeToast(o.id):o.gesture=void 0}})}updateTouchMove(e){this.queue.getElements().forEach(o=>{if(o.gesture){let l=e.clientX-o.gesture.touchStart.clientX;o.gesture.offsetX=l}})}updateToast(e,n){let o=this.queue.getElements(),l=o.findIndex(u=>u.id===e);l!==-1&&(o[l]=y(y({},o[l]),n),this.newToastBehaviorSubject.next([...o]))}static{this.\u0275fac=function(n){return new(n||t)(xe(Le))}}static{this.\u0275prov=B({token:t,factory:t.\u0275fac,providedIn:"root"})}}return t})();var yt=(t,i)=>i.id;function _t(t,i){if(t&1&&(s(0,"div",4),a(1),r()),t&2){let e=v().$implicit;d(),C(e.icon)}}function vt(t,i){t&1&&m(0,"div",11)}function Ct(t,i){t&1&&m(0,"div",12)}function bt(t,i){if(t&1&&(s(0,"div",6)(1,"strong",13),a(2),r()()),t&2){let e=v().$implicit;d(2),C(e.title)}}function St(t,i){if(t&1&&m(0,"div",7),t&2){let e=v().$implicit;c("innerHTML",e.message,De)}}function Et(t,i){if(t&1&&(s(0,"div",14),a(1),r()),t&2){let e=v().$implicit;d(),C(e.message)}}function Vt(t,i){if(t&1&&m(0,"div",15),t&2){let e=v().$implicit;de("transition-duration",e.duration,"ms"),w("animate",e.animatePbar)}}function xt(t,i){if(t&1){let e=U();s(0,"div",2)(1,"div",3),x(2,_t,2,1,"div",4)(3,vt,1,0)(4,Ct,1,0),s(5,"div",5),x(6,bt,3,1,"div",6)(7,St,1,1,"div",7)(8,Et,2,1),r(),s(9,"div",8)(10,"button",9),g("click",function(){let o=j(e).$implicit,l=v();return G(l.closeToast(o.id))}),a(11,"\xD7"),r()(),x(12,Vt,1,4,"div",10),r()()}if(t&2){let e=i.$implicit,n=v();de("transform","translateX("+((e.gesture==null?null:e.gesture.offsetX)||0)+"px)")("transition",e.gesture?"none":"transform 0.3s ease"),d(),w("fadeout",n.expires(e)),c("ngClass",e.type)("ngStyle",e.config==null?null:e.config.customStyle),ue("data-toast-id",e.id),d(),V(2,e.icon?2:e.config!=null&&e.config.loading?3:4),d(4),V(6,e.title?6:-1),d(),V(7,e.config!=null&&e.config.enableHtml?7:8),d(5),V(12,e.config!=null&&e.config.progressBar?12:-1)}}var pe=function(t){return t.TOP_LEFT="top-left",t.TOP="top",t.TOP_RIGHT="top-right",t.BOTTOM_LEFT="bottom-left",t.BOTTOM="bottom",t.BOTTOM_RIGHT="bottom-right",t}(pe||{}),Ue=(()=>{class t{constructor(e){this._toastService=e,this.position=pe.BOTTOM_RIGHT,this.duration=this._toastService.getDefaultDuration(),this.capacity=this._toastService.getCapacity(),this.toasts$=this._toastService.newToast$,this.positionMap={"top-left":["top","left"],top:["top","center"],"top-right":["top","right"],"bottom-left":["bottom","left"],bottom:["bottom","center"],"bottom-right":["bottom","right"]}}ngOnInit(){this._toastService.deleteToast$.subscribe(e=>{}),this._toastService.updateToast$.subscribe(e=>{})}ngOnChanges(e){e.duration&&this._toastService.setDefaultDuration(this.duration),e.capacity&&this._toastService.setCapacity(e.capacity.currentValue)}closeToast(e){this._toastService.closeToast(e)}containerClass(){return this.positionMap[this.position]??[pe.BOTTOM_RIGHT]}expires(e){return e.expires<Date.now()}getDurationUntilExpire(e){let n=e-Date.now();return n>0?n:0}onTouchStart(e){let n=this.localizeToastIdFromTouch(e);this._toastService.updateTouchStart(n,e.changedTouches[0])}onTouchEnd(e){this._toastService.updateTouchEnd(e.changedTouches[0])}onTouchMove(e){this._toastService.updateTouchMove(e.changedTouches[0])}localizeToastIdFromTouch(e){let n=e.target,o=4;for(;!n.classList.contains("toasty")&&o>0;)n=n.parentElement,o--;return parseInt(n.getAttribute("data-toast-id")||"0")}static{this.\u0275fac=function(n){return new(n||t)(p(k))}}static{this.\u0275cmp=M({type:t,selectors:[["toasty"]],hostBindings:function(n,o){n&1&&g("touchstart",function(u){return o.onTouchStart(u)})("touchend",function(u){return o.onTouchEnd(u)})("touchmove",function(u){return o.onTouchMove(u)})},inputs:{position:"position",duration:"duration",capacity:"capacity"},standalone:!0,features:[J,O],decls:4,vars:3,consts:[[1,"toasty-container",3,"ngClass"],[1,"toasty-wrapper",3,"transform","transition"],[1,"toasty-wrapper"],[1,"toasty",3,"ngClass","ngStyle"],[1,"toasty-icon"],[1,"toasty-content"],[1,"toasty-header"],[1,"toasty-body",3,"innerHTML"],[1,"toasty-close"],["type","button",1,"toasty-close",3,"click"],[1,"toasty-progress",3,"animate","transition-duration"],[1,"toasty-loading"],[1,"toasty-noicon"],[1,"toasty-title"],[1,"toasty-body"],[1,"toasty-progress"]],template:function(n,o){n&1&&(s(0,"div",0),Me(1,xt,13,13,"div",1,yt),Oe(3,"async"),r()),n&2&&(c("ngClass",o.containerClass()),d(),Ae(ke(3,1,o.toasts$)))},dependencies:[Ge,He,Re],styles:[".toasty-container[_ngcontent-%COMP%]{position:fixed;width:350px;padding:.5rem;right:0;bottom:0;pointer-events:none}.toasty-container.left[_ngcontent-%COMP%]{left:12px}.toasty-container.center[_ngcontent-%COMP%]{left:50%;transform:translate(-50%)}.toasty-container.right[_ngcontent-%COMP%]{right:12px}.toasty-container.top[_ngcontent-%COMP%]{top:10px;display:flex;flex-direction:column-reverse;justify-content:flex-end}.toasty-container.bottom[_ngcontent-%COMP%]{bottom:0}.toasty[_ngcontent-%COMP%]{padding:10px;border:1px solid rgba(0,0,0,.15);border-radius:8px;box-shadow:0 4px 10px #00000026;background-color:#fff;overflow:hidden;margin-bottom:1rem;font-family:Arial,Helvetica,sans-serif;animation-name:_ngcontent-%COMP%_toasty_fadein;animation-duration:.3s;animation-timing-function:ease-out;animation-delay:0s;animation-iteration-count:1;animation-fill-mode:forwards;display:flex;align-items:center;pointer-events:auto;touch-action:manipulation}@keyframes _ngcontent-%COMP%_toasty_fadein{0%{opacity:0;transform:translate(100%)}to{opacity:1;transform:translate(0)}}@keyframes _ngcontent-%COMP%_toasty_fadeout{0%{opacity:1}to{opacity:0;pointer-events:none}}.toasty.fadeout[_ngcontent-%COMP%]{animation-name:_ngcontent-%COMP%_toasty_fadeout;animation-duration:.5s;animation-timing-function:ease-in;animation-delay:0s;animation-iteration-count:1;animation-fill-mode:forwards;pointer-events:none}.toasty-icon[_ngcontent-%COMP%]{width:20px;height:20px;margin-right:.5rem;border-radius:4px}.toasty-loading[_ngcontent-%COMP%]{width:10px;height:10px;margin-right:.6rem;border-radius:50%;display:inline-block;border-top:3px solid;border-right:3px solid transparent;box-sizing:border-box;animation:_ngcontent-%COMP%_loading-toasty-anim 1s linear infinite}@keyframes _ngcontent-%COMP%_loading-toasty-anim{0%{transform:rotate(0)}to{transform:rotate(360deg)}}.toasty-noicon[_ngcontent-%COMP%]{width:7px}.toasty-content[_ngcontent-%COMP%]{flex-direction:column;flex:1}.toasty-header[_ngcontent-%COMP%]{display:flex;align-items:center}.toasty-title[_ngcontent-%COMP%]{font-size:.9rem;font-weight:700;margin-right:auto}.toasty-time[_ngcontent-%COMP%]{font-size:.85rem;color:#6c757d;margin-right:1rem}.toasty-close[_ngcontent-%COMP%]{background:none;border:none;font-size:1.4rem;line-height:1;cursor:pointer;color:inherit;opacity:.6}.toasty-body[_ngcontent-%COMP%]{font-size:.85rem;margin-top:1px}.toasty-close[_ngcontent-%COMP%]:hover{opacity:1}.toasty.none[_ngcontent-%COMP%]{background-color:#fff;border-color:#ededed;color:#171717}.toasty.success[_ngcontent-%COMP%]{background-color:#ecfdf3;border-color:#bffcd9;color:#008a2e}.toasty.error[_ngcontent-%COMP%]{background-color:#fff0f0;border-color:#ffe0e1;color:#e60000}.toasty.info[_ngcontent-%COMP%]{background-color:#f0f8ff;border-color:#dde7fd;color:#0973dc}.toasty.warning[_ngcontent-%COMP%]{background-color:#fffcf0;border-color:#fbeeb1;color:#dc7609}.toasty.custom[_ngcontent-%COMP%]{background-color:#fff}.toasty-progress[_ngcontent-%COMP%]{position:absolute;bottom:0;left:0;height:4px;width:100%;background-color:currentColor;opacity:.3;transform-origin:left;transform:scaleX(1);transition-property:transform;transition-timing-function:linear}.toasty-progress.animate[_ngcontent-%COMP%]{transform:scaleX(0)}"]})}}return t})();var Tt=["codeElement"];function Dt(t,i){if(t&1&&(s(0,"div",2)(1,"h4",7),a(2),r()()),t&2){let e=v();d(2),C(e.title)}}function Mt(t,i){if(t&1){let e=U();s(0,"button",8),g("click",function(){j(e);let o=v();return G(o.executeCode())}),a(1," Run "),r()}}var We=(()=>{class t{constructor(e){this._toastyService=e,this.title="",this.innerCode="",this.run=!0}ngAfterViewInit(){this.highlightCode()}highlightCode(){this.codeElement&&typeof Prism<"u"&&Prism.highlightElement(this.codeElement.nativeElement)}copy(){this._toastyService.showToast("Copied","",{type:h.Info})}executeCode(){try{new Function("_toastyService","ToastType",`
-        ${this.innerCode.replace(/this\._toastyService/g,"_toastyService")};
-      `)(this._toastyService,h)}catch(e){console.error("Error executing code:",e)}}static{this.\u0275fac=function(n){return new(n||t)(p(k))}}static{this.\u0275cmp=M({type:t,selectors:[["app-code"]],viewQuery:function(n,o){if(n&1&&Fe(Tt,5),n&2){let l;Ie(l=Ne())&&(o.codeElement=l.first)}},inputs:{title:"title",innerCode:"innerCode",run:"run"},standalone:!0,features:[O],decls:10,vars:3,consts:[["codeElement",""],[1,"codediv","relative"],[1,"bg-gray-200","px-4","py-2","border-b","border-gray-300","rounded-t"],[1,"bg-gray-100","text-gray-800","p-3","rounded","border","border-gray-300","mb-2"],[1,"text-sm","language-typescript"],["onclick","navigator.clipboard.writeText(this.parentElement.querySelector('code').innerText)",1,"cursor-pointer","absolute","top-2","right-2","px-2","py-1","text-xs","bg-gray-200","hover:bg-gray-300","rounded",3,"click"],[1,"absolute","bottom-2","right-2","px-2","py-1","text-xs","bg-blue-200","hover:bg-blue-300","rounded"],[1,"text-sm","font-semibold","text-gray-700","m-0"],[1,"absolute","bottom-2","right-2","px-2","py-1","text-xs","bg-blue-200","hover:bg-blue-300","rounded",3,"click"]],template:function(n,o){if(n&1){let l=U();s(0,"div",1),x(1,Dt,3,1,"div",2),s(2,"div",3)(3,"pre")(4,"code",4,0),a(6),r()(),s(7,"button",5),g("click",function(){return j(l),G(o.copy())}),a(8," Copy "),r(),x(9,Mt,2,0,"button",6),r()()}n&2&&(d(),V(1,o.title?1:-1),d(5),C(o.innerCode),d(3),V(9,o.run?9:-1))},styles:[".codediv[_ngcontent-%COMP%]{width:100%;overflow:auto;margin-top:4px}code[_ngcontent-%COMP%]{white-space:break-spaces}pre[_ngcontent-%COMP%]{padding:0;margin:0}"]})}}return t})();var Ke=(()=>{class t{constructor(e,n){this._renderer=e,this._elementRef=n,this.onChange=o=>{},this.onTouched=()=>{}}setProperty(e,n){this._renderer.setProperty(this._elementRef.nativeElement,e,n)}registerOnTouched(e){this.onTouched=e}registerOnChange(e){this.onChange=e}setDisabledState(e){this.setProperty("disabled",e)}static{this.\u0275fac=function(n){return new(n||t)(p(L),p(R))}}static{this.\u0275dir=_({type:t})}}return t})(),ie=(()=>{class t extends Ke{static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=H(t)))(o||t)}})()}static{this.\u0275dir=_({type:t,features:[E]})}}return t})(),Y=new D("");var At={provide:Y,useExisting:T(()=>ne),multi:!0};function Ft(){let t=he()?he().getUserAgent():"";return/android (\d+)/.test(t.toLowerCase())}var It=new D(""),ne=(()=>{class t extends Ke{constructor(e,n,o){super(e,n),this._compositionMode=o,this._composing=!1,this._compositionMode==null&&(this._compositionMode=!Ft())}writeValue(e){let n=e??"";this.setProperty("value",n)}_handleInput(e){(!this._compositionMode||this._compositionMode&&!this._composing)&&this.onChange(e)}_compositionStart(){this._composing=!0}_compositionEnd(e){this._composing=!1,this._compositionMode&&this.onChange(e)}static{this.\u0275fac=function(n){return new(n||t)(p(L),p(R),p(It,8))}}static{this.\u0275dir=_({type:t,selectors:[["input","formControlName","",3,"type","checkbox"],["textarea","formControlName",""],["input","formControl","",3,"type","checkbox"],["textarea","formControl",""],["input","ngModel","",3,"type","checkbox"],["textarea","ngModel",""],["","ngDefaultControl",""]],hostBindings:function(n,o){n&1&&g("input",function(u){return o._handleInput(u.target.value)})("blur",function(){return o.onTouched()})("compositionstart",function(){return o._compositionStart()})("compositionend",function(u){return o._compositionEnd(u.target.value)})},features:[N([At]),E]})}}return t})();var Nt=new D(""),Ot=new D("");function et(t){return t!=null}function tt(t){return Pe(t)?Se(t):t}function it(t){let i={};return t.forEach(e=>{i=e!=null?y(y({},i),e):i}),Object.keys(i).length===0?null:i}function nt(t,i){return i.map(e=>e(t))}function kt(t){return!t.validate}function ot(t){return t.map(i=>kt(i)?i:e=>i.validate(e))}function Pt(t){if(!t)return null;let i=t.filter(et);return i.length==0?null:function(e){return it(nt(e,i))}}function rt(t){return t!=null?Pt(ot(t)):null}function Bt(t){if(!t)return null;let i=t.filter(et);return i.length==0?null:function(e){let n=nt(e,i).map(tt);return Ve(n).pipe(Ee(it))}}function st(t){return t!=null?Bt(ot(t)):null}function qe(t,i){return t===null?[i]:Array.isArray(t)?[...t,i]:[t,i]}function jt(t){return t._rawValidators}function Gt(t){return t._rawAsyncValidators}function fe(t){return t?Array.isArray(t)?t:[t]:[]}function ee(t,i){return Array.isArray(t)?t.includes(i):t===i}function $e(t,i){let e=fe(i);return fe(t).forEach(o=>{ee(e,o)||e.push(o)}),e}function ze(t,i){return fe(i).filter(e=>!ee(t,e))}var te=class{constructor(){this._rawValidators=[],this._rawAsyncValidators=[],this._onDestroyCallbacks=[]}get value(){return this.control?this.control.value:null}get valid(){return this.control?this.control.valid:null}get invalid(){return this.control?this.control.invalid:null}get pending(){return this.control?this.control.pending:null}get disabled(){return this.control?this.control.disabled:null}get enabled(){return this.control?this.control.enabled:null}get errors(){return this.control?this.control.errors:null}get pristine(){return this.control?this.control.pristine:null}get dirty(){return this.control?this.control.dirty:null}get touched(){return this.control?this.control.touched:null}get status(){return this.control?this.control.status:null}get untouched(){return this.control?this.control.untouched:null}get statusChanges(){return this.control?this.control.statusChanges:null}get valueChanges(){return this.control?this.control.valueChanges:null}get path(){return null}_setValidators(i){this._rawValidators=i||[],this._composedValidatorFn=rt(this._rawValidators)}_setAsyncValidators(i){this._rawAsyncValidators=i||[],this._composedAsyncValidatorFn=st(this._rawAsyncValidators)}get validator(){return this._composedValidatorFn||null}get asyncValidator(){return this._composedAsyncValidatorFn||null}_registerOnDestroy(i){this._onDestroyCallbacks.push(i)}_invokeOnDestroyCallbacks(){this._onDestroyCallbacks.forEach(i=>i()),this._onDestroyCallbacks=[]}reset(i=void 0){this.control&&this.control.reset(i)}hasError(i,e){return this.control?this.control.hasError(i,e):!1}getError(i,e){return this.control?this.control.getError(i,e):null}},ge=class extends te{get formDirective(){return null}get path(){return null}},z=class extends te{constructor(){super(...arguments),this._parent=null,this.name=null,this.valueAccessor=null}},ye=class{constructor(i){this._cd=i}get isTouched(){return!!this._cd?.control?.touched}get isUntouched(){return!!this._cd?.control?.untouched}get isPristine(){return!!this._cd?.control?.pristine}get isDirty(){return!!this._cd?.control?.dirty}get isValid(){return!!this._cd?.control?.valid}get isInvalid(){return!!this._cd?.control?.invalid}get isPending(){return!!this._cd?.control?.pending}get isSubmitted(){return!!this._cd?.submitted}},Ht={"[class.ng-untouched]":"isUntouched","[class.ng-touched]":"isTouched","[class.ng-pristine]":"isPristine","[class.ng-dirty]":"isDirty","[class.ng-valid]":"isValid","[class.ng-invalid]":"isInvalid","[class.ng-pending]":"isPending"},ki=S(y({},Ht),{"[class.ng-submitted]":"isSubmitted"}),at=(()=>{class t extends ye{constructor(e){super(e)}static{this.\u0275fac=function(n){return new(n||t)(p(z,2))}}static{this.\u0275dir=_({type:t,selectors:[["","formControlName",""],["","ngModel",""],["","formControl",""]],hostVars:14,hostBindings:function(n,o){n&2&&w("ng-untouched",o.isUntouched)("ng-touched",o.isTouched)("ng-pristine",o.isPristine)("ng-dirty",o.isDirty)("ng-valid",o.isValid)("ng-invalid",o.isInvalid)("ng-pending",o.isPending)},features:[E]})}}return t})();var q="VALID",K="INVALID",P="PENDING",$="DISABLED";function Rt(t){return(oe(t)?t.validators:t)||null}function Lt(t){return Array.isArray(t)?rt(t):t||null}function Ut(t,i){return(oe(i)?i.asyncValidators:t)||null}function Wt(t){return Array.isArray(t)?st(t):t||null}function oe(t){return t!=null&&!Array.isArray(t)&&typeof t=="object"}var _e=class{constructor(i,e){this._pendingDirty=!1,this._hasOwnPendingAsyncValidator=!1,this._pendingTouched=!1,this._onCollectionChange=()=>{},this._parent=null,this.pristine=!0,this.touched=!1,this._onDisabledChange=[],this._assignValidators(i),this._assignAsyncValidators(e)}get validator(){return this._composedValidatorFn}set validator(i){this._rawValidators=this._composedValidatorFn=i}get asyncValidator(){return this._composedAsyncValidatorFn}set asyncValidator(i){this._rawAsyncValidators=this._composedAsyncValidatorFn=i}get parent(){return this._parent}get valid(){return this.status===q}get invalid(){return this.status===K}get pending(){return this.status==P}get disabled(){return this.status===$}get enabled(){return this.status!==$}get dirty(){return!this.pristine}get untouched(){return!this.touched}get updateOn(){return this._updateOn?this._updateOn:this.parent?this.parent.updateOn:"change"}setValidators(i){this._assignValidators(i)}setAsyncValidators(i){this._assignAsyncValidators(i)}addValidators(i){this.setValidators($e(i,this._rawValidators))}addAsyncValidators(i){this.setAsyncValidators($e(i,this._rawAsyncValidators))}removeValidators(i){this.setValidators(ze(i,this._rawValidators))}removeAsyncValidators(i){this.setAsyncValidators(ze(i,this._rawAsyncValidators))}hasValidator(i){return ee(this._rawValidators,i)}hasAsyncValidator(i){return ee(this._rawAsyncValidators,i)}clearValidators(){this.validator=null}clearAsyncValidators(){this.asyncValidator=null}markAsTouched(i={}){this.touched=!0,this._parent&&!i.onlySelf&&this._parent.markAsTouched(i)}markAllAsTouched(){this.markAsTouched({onlySelf:!0}),this._forEachChild(i=>i.markAllAsTouched())}markAsUntouched(i={}){this.touched=!1,this._pendingTouched=!1,this._forEachChild(e=>{e.markAsUntouched({onlySelf:!0})}),this._parent&&!i.onlySelf&&this._parent._updateTouched(i)}markAsDirty(i={}){this.pristine=!1,this._parent&&!i.onlySelf&&this._parent.markAsDirty(i)}markAsPristine(i={}){this.pristine=!0,this._pendingDirty=!1,this._forEachChild(e=>{e.markAsPristine({onlySelf:!0})}),this._parent&&!i.onlySelf&&this._parent._updatePristine(i)}markAsPending(i={}){this.status=P,i.emitEvent!==!1&&this.statusChanges.emit(this.status),this._parent&&!i.onlySelf&&this._parent.markAsPending(i)}disable(i={}){let e=this._parentMarkedDirty(i.onlySelf);this.status=$,this.errors=null,this._forEachChild(n=>{n.disable(S(y({},i),{onlySelf:!0}))}),this._updateValue(),i.emitEvent!==!1&&(this.valueChanges.emit(this.value),this.statusChanges.emit(this.status)),this._updateAncestors(S(y({},i),{skipPristineCheck:e})),this._onDisabledChange.forEach(n=>n(!0))}enable(i={}){let e=this._parentMarkedDirty(i.onlySelf);this.status=q,this._forEachChild(n=>{n.enable(S(y({},i),{onlySelf:!0}))}),this.updateValueAndValidity({onlySelf:!0,emitEvent:i.emitEvent}),this._updateAncestors(S(y({},i),{skipPristineCheck:e})),this._onDisabledChange.forEach(n=>n(!1))}_updateAncestors(i){this._parent&&!i.onlySelf&&(this._parent.updateValueAndValidity(i),i.skipPristineCheck||this._parent._updatePristine(),this._parent._updateTouched())}setParent(i){this._parent=i}getRawValue(){return this.value}updateValueAndValidity(i={}){this._setInitialStatus(),this._updateValue(),this.enabled&&(this._cancelExistingSubscription(),this.errors=this._runValidator(),this.status=this._calculateStatus(),(this.status===q||this.status===P)&&this._runAsyncValidator(i.emitEvent)),i.emitEvent!==!1&&(this.valueChanges.emit(this.value),this.statusChanges.emit(this.status)),this._parent&&!i.onlySelf&&this._parent.updateValueAndValidity(i)}_updateTreeValidity(i={emitEvent:!0}){this._forEachChild(e=>e._updateTreeValidity(i)),this.updateValueAndValidity({onlySelf:!0,emitEvent:i.emitEvent})}_setInitialStatus(){this.status=this._allControlsDisabled()?$:q}_runValidator(){return this.validator?this.validator(this):null}_runAsyncValidator(i){if(this.asyncValidator){this.status=P,this._hasOwnPendingAsyncValidator=!0;let e=tt(this.asyncValidator(this));this._asyncValidationSubscription=e.subscribe(n=>{this._hasOwnPendingAsyncValidator=!1,this.setErrors(n,{emitEvent:i})})}}_cancelExistingSubscription(){this._asyncValidationSubscription&&(this._asyncValidationSubscription.unsubscribe(),this._hasOwnPendingAsyncValidator=!1)}setErrors(i,e={}){this.errors=i,this._updateControlsErrors(e.emitEvent!==!1)}get(i){let e=i;return e==null||(Array.isArray(e)||(e=e.split(".")),e.length===0)?null:e.reduce((n,o)=>n&&n._find(o),this)}getError(i,e){let n=e?this.get(e):this;return n&&n.errors?n.errors[i]:null}hasError(i,e){return!!this.getError(i,e)}get root(){let i=this;for(;i._parent;)i=i._parent;return i}_updateControlsErrors(i){this.status=this._calculateStatus(),i&&this.statusChanges.emit(this.status),this._parent&&this._parent._updateControlsErrors(i)}_initObservables(){this.valueChanges=new Q,this.statusChanges=new Q}_calculateStatus(){return this._allControlsDisabled()?$:this.errors?K:this._hasOwnPendingAsyncValidator||this._anyControlsHaveStatus(P)?P:this._anyControlsHaveStatus(K)?K:q}_anyControlsHaveStatus(i){return this._anyControls(e=>e.status===i)}_anyControlsDirty(){return this._anyControls(i=>i.dirty)}_anyControlsTouched(){return this._anyControls(i=>i.touched)}_updatePristine(i={}){this.pristine=!this._anyControlsDirty(),this._parent&&!i.onlySelf&&this._parent._updatePristine(i)}_updateTouched(i={}){this.touched=this._anyControlsTouched(),this._parent&&!i.onlySelf&&this._parent._updateTouched(i)}_registerOnCollectionChange(i){this._onCollectionChange=i}_setUpdateStrategy(i){oe(i)&&i.updateOn!=null&&(this._updateOn=i.updateOn)}_parentMarkedDirty(i){let e=this._parent&&this._parent.dirty;return!i&&!!e&&!this._parent._anyControlsDirty()}_find(i){return null}_assignValidators(i){this._rawValidators=Array.isArray(i)?i.slice():i,this._composedValidatorFn=Lt(this._rawValidators)}_assignAsyncValidators(i){this._rawAsyncValidators=Array.isArray(i)?i.slice():i,this._composedAsyncValidatorFn=Wt(this._rawAsyncValidators)}};var lt=new D("CallSetDisabledState",{providedIn:"root",factory:()=>ve}),ve="always";function qt(t,i){return[...i.path,t]}function $t(t,i,e=ve){Yt(t,i),i.valueAccessor.writeValue(t.value),(t.disabled||e==="always")&&i.valueAccessor.setDisabledState?.(t.disabled),Xt(t,i),Jt(t,i),Zt(t,i),zt(t,i)}function Ye(t,i){t.forEach(e=>{e.registerOnValidatorChange&&e.registerOnValidatorChange(i)})}function zt(t,i){if(i.valueAccessor.setDisabledState){let e=n=>{i.valueAccessor.setDisabledState(n)};t.registerOnDisabledChange(e),i._registerOnDestroy(()=>{t._unregisterOnDisabledChange(e)})}}function Yt(t,i){let e=jt(t);i.validator!==null?t.setValidators(qe(e,i.validator)):typeof e=="function"&&t.setValidators([e]);let n=Gt(t);i.asyncValidator!==null?t.setAsyncValidators(qe(n,i.asyncValidator)):typeof n=="function"&&t.setAsyncValidators([n]);let o=()=>t.updateValueAndValidity();Ye(i._rawValidators,o),Ye(i._rawAsyncValidators,o)}function Xt(t,i){i.valueAccessor.registerOnChange(e=>{t._pendingValue=e,t._pendingChange=!0,t._pendingDirty=!0,t.updateOn==="change"&&ut(t,i)})}function Zt(t,i){i.valueAccessor.registerOnTouched(()=>{t._pendingTouched=!0,t.updateOn==="blur"&&t._pendingChange&&ut(t,i),t.updateOn!=="submit"&&t.markAsTouched()})}function ut(t,i){t._pendingDirty&&t.markAsDirty(),t.setValue(t._pendingValue,{emitModelToViewChange:!1}),i.viewToModelUpdate(t._pendingValue),t._pendingChange=!1}function Jt(t,i){let e=(n,o)=>{i.valueAccessor.writeValue(n),o&&i.viewToModelUpdate(n)};t.registerOnChange(e),i._registerOnDestroy(()=>{t._unregisterOnChange(e)})}function Qt(t,i){if(!t.hasOwnProperty("model"))return!1;let e=t.model;return e.isFirstChange()?!0:!Object.is(i,e.currentValue)}function Kt(t){return Object.getPrototypeOf(t.constructor)===ie}function ei(t,i){if(!i)return null;Array.isArray(i);let e,n,o;return i.forEach(l=>{l.constructor===ne?e=l:Kt(l)?n=l:o=l}),o||n||e||null}function Xe(t,i){let e=t.indexOf(i);e>-1&&t.splice(e,1)}function Ze(t){return typeof t=="object"&&t!==null&&Object.keys(t).length===2&&"value"in t&&"disabled"in t}var ti=class extends _e{constructor(i=null,e,n){super(Rt(e),Ut(n,e)),this.defaultValue=null,this._onChange=[],this._pendingChange=!1,this._applyFormState(i),this._setUpdateStrategy(e),this._initObservables(),this.updateValueAndValidity({onlySelf:!0,emitEvent:!!this.asyncValidator}),oe(e)&&(e.nonNullable||e.initialValueIsDefault)&&(Ze(i)?this.defaultValue=i.value:this.defaultValue=i)}setValue(i,e={}){this.value=this._pendingValue=i,this._onChange.length&&e.emitModelToViewChange!==!1&&this._onChange.forEach(n=>n(this.value,e.emitViewToModelChange!==!1)),this.updateValueAndValidity(e)}patchValue(i,e={}){this.setValue(i,e)}reset(i=this.defaultValue,e={}){this._applyFormState(i),this.markAsPristine(e),this.markAsUntouched(e),this.setValue(this.value,e),this._pendingChange=!1}_updateValue(){}_anyControls(i){return!1}_allControlsDisabled(){return this.disabled}registerOnChange(i){this._onChange.push(i)}_unregisterOnChange(i){Xe(this._onChange,i)}registerOnDisabledChange(i){this._onDisabledChange.push(i)}_unregisterOnDisabledChange(i){Xe(this._onDisabledChange,i)}_forEachChild(i){}_syncPendingControls(){return this.updateOn==="submit"&&(this._pendingDirty&&this.markAsDirty(),this._pendingTouched&&this.markAsTouched(),this._pendingChange)?(this.setValue(this._pendingValue,{onlySelf:!0,emitModelToViewChange:!1}),!0):!1}_applyFormState(i){Ze(i)?(this.value=this._pendingValue=i.value,i.disabled?this.disable({onlySelf:!0,emitEvent:!1}):this.enable({onlySelf:!0,emitEvent:!1})):this.value=this._pendingValue=i}};var ii={provide:z,useExisting:T(()=>Ce)},Je=Promise.resolve(),Ce=(()=>{class t extends z{constructor(e,n,o,l,u,f){super(),this._changeDetectorRef=u,this.callSetDisabledState=f,this.control=new ti,this._registered=!1,this.name="",this.update=new Q,this._parent=e,this._setValidators(n),this._setAsyncValidators(o),this.valueAccessor=ei(this,l)}ngOnChanges(e){if(this._checkForErrors(),!this._registered||"name"in e){if(this._registered&&(this._checkName(),this.formDirective)){let n=e.name.previousValue;this.formDirective.removeControl({name:n,path:this._getPath(n)})}this._setUpControl()}"isDisabled"in e&&this._updateDisabled(e),Qt(e,this.viewModel)&&(this._updateValue(this.model),this.viewModel=this.model)}ngOnDestroy(){this.formDirective&&this.formDirective.removeControl(this)}get path(){return this._getPath(this.name)}get formDirective(){return this._parent?this._parent.formDirective:null}viewToModelUpdate(e){this.viewModel=e,this.update.emit(e)}_setUpControl(){this._setUpdateStrategy(),this._isStandalone()?this._setUpStandalone():this.formDirective.addControl(this),this._registered=!0}_setUpdateStrategy(){this.options&&this.options.updateOn!=null&&(this.control._updateOn=this.options.updateOn)}_isStandalone(){return!this._parent||!!(this.options&&this.options.standalone)}_setUpStandalone(){$t(this.control,this,this.callSetDisabledState),this.control.updateValueAndValidity({emitEvent:!1})}_checkForErrors(){this._isStandalone()||this._checkParentType(),this._checkName()}_checkParentType(){}_checkName(){this.options&&this.options.name&&(this.name=this.options.name),!this._isStandalone()&&this.name}_updateValue(e){Je.then(()=>{this.control.setValue(e,{emitViewToModelChange:!1}),this._changeDetectorRef?.markForCheck()})}_updateDisabled(e){let n=e.isDisabled.currentValue,o=n!==0&&je(n);Je.then(()=>{o&&!this.control.disabled?this.control.disable():!o&&this.control.disabled&&this.control.enable(),this._changeDetectorRef?.markForCheck()})}_getPath(e){return this._parent?qt(e,this._parent):[e]}static{this.\u0275fac=function(n){return new(n||t)(p(ge,9),p(Nt,10),p(Ot,10),p(Y,10),p(Be,8),p(lt,8))}}static{this.\u0275dir=_({type:t,selectors:[["","ngModel","",3,"formControlName","",3,"formControl",""]],inputs:{name:"name",isDisabled:[Z.None,"disabled","isDisabled"],model:[Z.None,"ngModel","model"],options:[Z.None,"ngModelOptions","options"]},outputs:{update:"ngModelChange"},exportAs:["ngModel"],features:[N([ii]),E,J]})}}return t})();var ni={provide:Y,useExisting:T(()=>be),multi:!0},be=(()=>{class t extends ie{writeValue(e){this.setProperty("value",parseFloat(e))}registerOnChange(e){this.onChange=n=>{e(n==""?null:parseFloat(n))}}static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=H(t)))(o||t)}})()}static{this.\u0275dir=_({type:t,selectors:[["input","type","range","formControlName",""],["input","type","range","formControl",""],["input","type","range","ngModel",""]],hostBindings:function(n,o){n&1&&g("change",function(u){return o.onChange(u.target.value)})("input",function(u){return o.onChange(u.target.value)})("blur",function(){return o.onTouched()})},features:[N([ni]),E]})}}return t})();var oi={provide:Y,useExisting:T(()=>re),multi:!0};function dt(t,i){return t==null?`${i}`:(i&&typeof i=="object"&&(i="Object"),`${t}: ${i}`.slice(0,50))}function ri(t){return t.split(":")[0]}var re=(()=>{class t extends ie{constructor(){super(...arguments),this._optionMap=new Map,this._idCounter=0,this._compareWith=Object.is}set compareWith(e){this._compareWith=e}writeValue(e){this.value=e;let n=this._getOptionId(e),o=dt(n,e);this.setProperty("value",o)}registerOnChange(e){this.onChange=n=>{this.value=this._getOptionValue(n),e(this.value)}}_registerOption(){return(this._idCounter++).toString()}_getOptionId(e){for(let n of this._optionMap.keys())if(this._compareWith(this._optionMap.get(n),e))return n;return null}_getOptionValue(e){let n=ri(e);return this._optionMap.has(n)?this._optionMap.get(n):e}static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=H(t)))(o||t)}})()}static{this.\u0275dir=_({type:t,selectors:[["select","formControlName","",3,"multiple",""],["select","formControl","",3,"multiple",""],["select","ngModel","",3,"multiple",""]],hostBindings:function(n,o){n&1&&g("change",function(u){return o.onChange(u.target.value)})("blur",function(){return o.onTouched()})},inputs:{compareWith:"compareWith"},features:[N([oi]),E]})}}return t})(),ct=(()=>{class t{constructor(e,n,o){this._element=e,this._renderer=n,this._select=o,this._select&&(this.id=this._select._registerOption())}set ngValue(e){this._select!=null&&(this._select._optionMap.set(this.id,e),this._setElementValue(dt(this.id,e)),this._select.writeValue(this._select.value))}set value(e){this._setElementValue(e),this._select&&this._select.writeValue(this._select.value)}_setElementValue(e){this._renderer.setProperty(this._element.nativeElement,"value",e)}ngOnDestroy(){this._select&&(this._select._optionMap.delete(this.id),this._select.writeValue(this._select.value))}static{this.\u0275fac=function(n){return new(n||t)(p(R),p(L),p(re,9))}}static{this.\u0275dir=_({type:t,selectors:[["option"]],inputs:{ngValue:"ngValue",value:"value"}})}}return t})(),si={provide:Y,useExisting:T(()=>ht),multi:!0};function Qe(t,i){return t==null?`${i}`:(typeof i=="string"&&(i=`'${i}'`),i&&typeof i=="object"&&(i="Object"),`${t}: ${i}`.slice(0,50))}function ai(t){return t.split(":")[0]}var ht=(()=>{class t extends ie{constructor(){super(...arguments),this._optionMap=new Map,this._idCounter=0,this._compareWith=Object.is}set compareWith(e){this._compareWith=e}writeValue(e){this.value=e;let n;if(Array.isArray(e)){let o=e.map(l=>this._getOptionId(l));n=(l,u)=>{l._setSelected(o.indexOf(u.toString())>-1)}}else n=(o,l)=>{o._setSelected(!1)};this._optionMap.forEach(n)}registerOnChange(e){this.onChange=n=>{let o=[],l=n.selectedOptions;if(l!==void 0){let u=l;for(let f=0;f<u.length;f++){let b=u[f],se=this._getOptionValue(b.value);o.push(se)}}else{let u=n.options;for(let f=0;f<u.length;f++){let b=u[f];if(b.selected){let se=this._getOptionValue(b.value);o.push(se)}}}this.value=o,e(o)}}_registerOption(e){let n=(this._idCounter++).toString();return this._optionMap.set(n,e),n}_getOptionId(e){for(let n of this._optionMap.keys())if(this._compareWith(this._optionMap.get(n)._value,e))return n;return null}_getOptionValue(e){let n=ai(e);return this._optionMap.has(n)?this._optionMap.get(n)._value:e}static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=H(t)))(o||t)}})()}static{this.\u0275dir=_({type:t,selectors:[["select","multiple","","formControlName",""],["select","multiple","","formControl",""],["select","multiple","","ngModel",""]],hostBindings:function(n,o){n&1&&g("change",function(u){return o.onChange(u.target)})("blur",function(){return o.onTouched()})},inputs:{compareWith:"compareWith"},features:[N([si]),E]})}}return t})(),mt=(()=>{class t{constructor(e,n,o){this._element=e,this._renderer=n,this._select=o,this._select&&(this.id=this._select._registerOption(this))}set ngValue(e){this._select!=null&&(this._value=e,this._setElementValue(Qe(this.id,e)),this._select.writeValue(this._select.value))}set value(e){this._select?(this._value=e,this._setElementValue(Qe(this.id,e)),this._select.writeValue(this._select.value)):this._setElementValue(e)}_setElementValue(e){this._renderer.setProperty(this._element.nativeElement,"value",e)}_setSelected(e){this._renderer.setProperty(this._element.nativeElement,"selected",e)}ngOnDestroy(){this._select&&(this._select._optionMap.delete(this.id),this._select.writeValue(this._select.value))}static{this.\u0275fac=function(n){return new(n||t)(p(R),p(L),p(ht,9))}}static{this.\u0275dir=_({type:t,selectors:[["option"]],inputs:{ngValue:"ngValue",value:"value"}})}}return t})();var li=(()=>{class t{static{this.\u0275fac=function(n){return new(n||t)}}static{this.\u0275mod=le({type:t})}static{this.\u0275inj=ae({})}}return t})();var pt=(()=>{class t{static withConfig(e){return{ngModule:t,providers:[{provide:lt,useValue:e.callSetDisabledState??ve}]}}static{this.\u0275fac=function(n){return new(n||t)}}static{this.\u0275mod=le({type:t})}static{this.\u0275inj=ae({imports:[li]})}}return t})();var Ri=(()=>{class t{constructor(e){this._toastyService=e,this.title="",this.message="",this.containerDuration=5e3,this.containerPosition="bottom-right",this.containerCapacity=10,this.durationToast=5,this.libName="ng-toast-notify",this.sidebarOpen=!1,this.toastMessages=[{title:"Mission Accomplished!",message:"You pressed a button. NASA is impressed."},{title:"Toast Success!",message:"This notification is brought to you by JavaScript and caffeine."},{title:"Brain.exe has stopped",message:"Just kidding! Everything\u2019s fine. Probably."},{title:"You\u2019ve Got Mail!",message:"Just kidding. It\u2019s 2025. No one sends emails anymore."},{title:"Bug? Never Heard of Her",message:"That action definitely didn\u2019t break anything. Maybe."},{title:"Toast Delivered",message:"Served fresh and slightly over-engineered."},{title:"Cool Move",message:"That click had style. Hollywood is calling."},{title:"Clean Slate",message:"We wiped it. Like, really wiped it. Hope that was okay."},{title:"Access Granted",message:"You are now officially allowed to feel powerful."},{title:"Fancy That!",message:"A toast so elegant, it sips its coffee with its pinky up."},{title:"Action Completed",message:"Your changes have been saved successfully."},{title:"Data Submitted",message:"The form was sent and processed correctly."},{title:"Warning",message:"Please check the required fields before continuing."},{title:"Error Occurred",message:"Something went wrong. Try again or contact support."},{title:"Changes Updated",message:"Your settings have been updated."},{title:"Logged In",message:"Welcome back! You have successfully signed in."},{title:"Access Denied",message:"You don\u2019t have permission to perform this action."},{title:"Session Expired",message:"Please log in again to continue."},{title:"Item Deleted",message:"The selected item has been removed from the list."},{title:"Autosaved",message:"Your progress was saved automatically."}],this.infosuccessetctoasts=`
-  this._toastyService.showToast('Info', 'Information body content', { type: ToastType.Info})
-  this._toastyService.showToast('Success', 'Success body content', { type: ToastType.Success})
-  this._toastyService.showToast('Warning', 'Warning body content', { type: ToastType.Warning})
-  this._toastyService.showToast('Error', 'Error body content', { type: ToastType.Error})
-  `,this.componentcode=`
-  import { ToastType, ToastyComponent, ToastyPromise, ToastyService } from '${this.libName}'; 
-  @Component({
-    selector: 'app-root',
-    standalone: true,
-    imports: [ToastyComponent],
-    templateUrl: './app.component.html',
-    styleUrl: './app.component.css'
-  })
-  `,this.customcode=`
-  const style = {
-      "background-image": "linear-gradient(to right, #e52d27 0%, #b31217  51%, #e52d27  100%)",
-      "color": "#fff"
-    };
-    this._toastyService.showToast('Custom', 'Some body content', { type: ToastType.Custom, customStyle: style });
-  `,this.customcode2=`
-  const style = {
-      "background-image": "linear-gradient(220.55deg, #FFE70B 0%, #27B643 100%)",
-      "border": "1px solid rgb(182 214 19)",
-      "color": "#fff"
-    };
-    this._toastyService.showToast('Custom', 'Some body content', { type: ToastType.Custom, customStyle: style });
-  `,this.customcode3=`
-    const style = { 
-      "gap": "12px",
-      "padding": "14px 28px",
-      "background": "linear-gradient(145deg, #0f0f0f, #1c1c1c)",
-      "border": "2px solid rgba(255, 255, 255, 0.2)",
-      "border-radius": "100px",
-      "color": "#fff",
-      "font-size": "16px",
-      "font-weight": "600",
-      "letter-spacing": "0.5px",
-      "cursor": "pointer",
-      "overflow": "hidden",
-      "box-shadow": "0 0 20px rgba(0, 255, 255, 0.1)",
-      "backdrop-filter": "blur(8px)"
-    };
-    this._toastyService.showToast('Custom', 'Some body content', { type: ToastType.Custom, customStyle: style });
-  `,this.longToast=`
-  this._toastyService.showToast("Long text", "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla tincidunt euismod nunc quis suscipit. Curabitur sed semper tortor, at efficitur leo. Sed justo nulla, ullamcorper sed maximus a, vehicula ut nulla. Vivamus lobortis lacinia mi eu ultricies. Suspendisse quis sem ante. Vivamus consectetur lectus eget quam venenatis, sagittis aliquam lectus viverra. Donec nec molestie magna, ut ultricies tellus. Praesent faucibus urna imperdiet vulputate dapibus. Sed sit amet dolor odio. Donec est nulla, cursus sed aliquam at, cursus sit amet turpis. Nulla ut dolor eget dui laoreet hendrerit vel at tellus. Nullam aliquet sem at mauris scelerisque molestie. Nullam imperdiet blandit est et aliquam. Vestibulum euismod, libero id tincidunt rutrum, nulla augue aliquet risus, a rutrum elit orci eget lectus.");
-  `,this.loadingToast=`
-  this._toastyService.showToast('Wait', 'Process may take a few seconds to complete', { type: ToastType.Custom, loading:true });
-  `,this.promiseToast=`
-    simulateApiCall(): Promise<{ name: string }> {
-      return new Promise((resolve, reject) =>
-        setTimeout(
-            () => Math.random() < 0.5 ? resolve({ name: 'Toasty resolved' }) : reject(new Error('Toasty rejected'))
-        , 2000));
-    }
-
-    this._toastyService.showToastPromise('Fetching API', 
-                                          {
-                                            promise: this.simulateApiCall(),
-                                            loading: 'Loading',
-                                            success: (data) => \`Ok \${data.name}\`,
-                                            error: 'Error'
-                                          });  
-  `}basicToast(){this._toastyService.showToast("Title","Message body")}infoToast(){this._toastyService.showToast("Info","Information body content",{type:h.Info})}successToast(){this._toastyService.showToast("Success","Success body content",{type:h.Success})}WarningToast(){this._toastyService.showToast("Warning","Warning body content",{type:h.Warning})}ErrorToast(){this._toastyService.showToast("Error","Error body content",{type:h.Error})}htmlToast(){this._toastyService.showToast("Html","<b>Details</b><ul><li>item 1</li><li>item 2</li></ul>",{type:h.Basic,enableHtml:!0})}htmlToast2(){this._toastyService.showToast("Html","Data successfully saved. <img width=80 src=./assets/img/dancing-spiderman.gif /> ",{type:h.Success,enableHtml:!0})}htmlToast3(){this._toastyService.showToast("Html","Achievement unlocked. <img width=100 src=./assets/img/wizard.gif /> ",{type:h.Info,enableHtml:!0})}customStyleToast(){let e={"background-image":"linear-gradient(to right, #e52d27 0%, #b31217  51%, #e52d27  100%)",color:"#fff"};this._toastyService.showToast("Custom","Some body content",{type:h.Custom,customStyle:e})}emojiToast(){this._toastyService.showToast("Error","\u2665\uFE0Flove emoji toast\u2665\uFE0F")}makeToast(){let e=[h.Info,h.Success,h.Warning,h.Basic,h.Error],n=e[Math.floor(Math.random()*e.length)],o=Math.floor(Math.random()*this.toastMessages.length),l=[{"background-image":"linear-gradient(to right, #77A1D3 0%, #79CBCA  51%, #77A1D3  100%)",color:"#fff","font-family":"cursive"},{"background-image":"linear-gradient(to right, #e52d27 0%, #b31217  51%, #e52d27  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, #FF512F 0%, #F09819  51%, #FF512F  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, #3D7EAA 0%, #FFE47A  51%, #3D7EAA  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, #BA8B02 0%, #181818  51%, #BA8B02  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, rgb(59, 59, 59), rgb(34, 34, 34))",color:"#fff"},{"background-image":"linear-gradient(97deg, #7918c8, #f900ff)",color:"#fff"},{"background-image":"linear-gradient(220.55deg, #FFE70B 0%, #27B643 100%)",border:"1px solid rgb(182 214 19)",color:"#fff"}],u=l[Math.floor(Math.random()*l.length)];this._toastyService.showToast(this.toastMessages[o].title,this.toastMessages[o].message,{type:h.Custom,enableHtml:!0,customStyle:u})}randomToast(){let e=[this.basicToast,this.infoToast,this.successToast,this.WarningToast,this.ErrorToast,this.makeToast,this.htmlToast,this.htmlToast2,this.htmlToast3,this.customStyleToast,this.clickLongToast,this.clickLoadingToast,this.clickPromiseToast,this.clickEmojiToast,this.clickBeepToast,this.clickProgressBarToast];e[Math.floor(Math.random()*e.length)].bind(this)()}clickProgressBarToast(){this._toastyService.showToast("Progress","This toast has a progress bar",{type:h.Info,progressBar:!0})}clickBeepToast(){this._toastyService.showToast("Beep","This toast goes beep beep!",{type:h.Info,beep:!0})}clickLongToast(){this._toastyService.showToast("Long text","Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla tincidunt euismod nunc quis suscipit. Curabitur sed semper tortor, at efficitur leo. Sed justo nulla, ullamcorper sed maximus a, vehicula ut nulla. Vivamus lobortis lacinia mi eu ultricies. Suspendisse quis sem ante. Vivamus consectetur lectus eget quam venenatis, sagittis aliquam lectus viverra. Donec nec molestie magna, ut ultricies tellus. Praesent faucibus urna imperdiet vulputate dapibus. Sed sit amet dolor odio. Donec est nulla, cursus sed aliquam at, cursus sit amet turpis. Nulla ut dolor eget dui laoreet hendrerit vel at tellus. Nullam aliquet sem at mauris scelerisque molestie. Nullam imperdiet blandit est et aliquam. Vestibulum euismod, libero id tincidunt rutrum, nulla augue aliquet risus, a rutrum elit orci eget lectus.")}clickLoadingToast(){this._toastyService.showToast("Wait","Process may take a few seconds to complete",{type:h.Custom,loading:!0})}clickEmojiToast(){this._toastyService.showToast("Error","\u2665\uFE0Flove emoji toast\u2665\uFE0F")}clickPromiseToast(){this._toastyService.showToastPromise("Fetching API",{promise:this.simulateApiCall(),loading:"Loading",success:e=>`Ok ${e.name}`,error:"Error"})}simulateApiCall(){return new Promise((e,n)=>setTimeout(()=>Math.random()<.5?e({name:"Toasty resolved"}):n(new Error("Toasty rejected")),2e3))}static{this.\u0275fac=function(n){return new(n||t)(p(k))}}static{this.\u0275cmp=M({type:t,selectors:[["app-ng-toast-notify"]],standalone:!0,features:[O],decls:319,vars:57,consts:[[1,"min-h-screen","bg-white","text-gray-900","flex"],["id","menu-toggle",1,"md:hidden","p-4","focus:outline-none","absolute","top-2","right-2","z-50",3,"click"],["xmlns","http://www.w3.org/2000/svg","fill","none","viewBox","0 0 24 24","stroke","currentColor",1,"h-6","w-6","text-gray-800"],["stroke-linecap","round","stroke-linejoin","round","stroke-width","2","d","M4 6h16M4 12h16m-7 6h7"],[1,"md:w-256","border-r","border-gray-200","h-screen","sticky","top-0","p-6","space-y-6","text-sm","overflow-y-auto","md:translate-x-0","transition-transform","duration-300"],[1,"text-xs","font-semibold","text-gray-400","uppercase","mb-2"],[1,"space-y-1"],["href","#install",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#addcomponent",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#useservice",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#basic-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#info-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#html-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#custom-css-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#long-message-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#loading-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#promise-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#emoji-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#swipe",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#beep",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#progressbar",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#sticky",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#container-settings",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],["href","#demo-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100"],[1,"px-8","py-10","space-y-5","font-sans","text-sm"],[1,"text-center","space-y-4","w-5/6"],[1,"text-5xl","font-bold","tracking-tight"],[1,"text-lg","text-gray-600"],[1,"flex","gap-4","justify-center"],[1,"cursor-pointer","px-5","py-2","bg-emerald-500","hover:bg-emerald-600","text-white","rounded","shadow","font-medium",3,"click"],["href","https://github.com/rickmarina/ng-toast-notify","target","_blank",1,"px-5","py-2","border","border-gray-300","rounded","hover:bg-gray-100","transition","text-sm"],["id","install",1,"w-full","mx-auto"],[1,"text-2xl","font-semibold","mb-2"],[1,"block","text-sm","font-medium","text-gray-700","mb-2"],[3,"innerCode","run"],["id","addcomponent",1,"w-full","mx-auto"],[1,"block","text-sm","text-gray-700","mb-2"],["title","component.html",3,"innerCode","run"],[1,""],["href","#container-settings",1,"font-medium","text-blue-600","dark:text-blue-500","hover:underline"],[1,"w-full","mx-auto","my-1"],["title","component.ts",3,"innerCode","run"],["id","useservice",1,"w-full","mx-auto"],[1,"block","mb-2"],["title","Inject the service in your component",3,"innerCode","run"],["title","Use service to display toasts",3,"innerCode","run"],["id","basic-toast",1,"w-full","mx-auto"],["title","Basic toast",3,"innerCode","run"],["id","info-toast",1,"w-full","mx-auto"],[1,"b",3,"click"],[3,"innerCode"],["id","html-toast",1,"w-full","mx-auto"],[1,"text-sm","text-gray-600","mb-4"],[3,"title","innerCode"],["id","custom-css-toast",1,"w-full","max-auto"],["title","Custom red toast",3,"innerCode"],["title","Mango toast",3,"innerCode"],["title","Blackhole toast",3,"innerCode"],["id","long-message-toast",1,"w-full","max-auto"],["title","Long message toast",3,"innerCode"],["id","loading-toast",1,"w-full","max-auto"],["title","Loading",3,"innerCode"],["id","promise-toast",1,"w-full","max-auto"],[1,"w-4/6"],["title","Promise",3,"innerCode","run"],["id","emoji-toast",1,"w-full","max-auto"],["title","Emoji",3,"innerCode"],["id","swipe",1,"w-full","max-auto"],["title","Swipe to close",3,"innerCode","run"],["id","beep",1,"w-full","max-auto"],["title","Notify sound",3,"innerCode","run"],["id","progressbar",1,"w-full","max-auto"],["title","Progress bar",3,"innerCode","run"],["id","sticky",1,"w-full","max-auto"],["title","Sticky toast",3,"innerCode","run"],["id","container-settings",1,"w-full","max-auto"],["title","Position",3,"innerCode","run"],["title","Duration",3,"innerCode","run"],["title","All settings together",3,"innerCode","run"],[1,"w-5/6"],[1,"font-medium"],[1,"overflow-x-auto"],[1,"min-w-full","text-sm","text-left","border","border-gray-200","rounded","overflow-hidden"],[1,"bg-gray-100","text-xs","uppercase","text-gray-600","border-b","border-gray-200"],[1,"px-4","py-2"],[1,"divide-y","divide-gray-200"],[1,"px-4","py-2","font-medium","text-gray-900"],[1,"px-4","py-2","text-gray-700"],[1,"px-4","py-2","text-gray-500"],["id","demo-toast",1,"w-full","space-y-4","mx-auto","hidden"],[1,"p-3","border","border-gray-200","rounded","space-y-3"],[1,"text-base","font-semibold"],[1,"flex","gap-2"],[1,"flex-1"],[1,"block","text-xs","font-medium","text-gray-700","mb-1"],["type","text","placeholder","Title",1,"w-full","px-3","py-1.5","border","border-gray-300","rounded",3,"ngModelChange","ngModel"],["rows","2","placeholder","Message",1,"w-full","px-3","py-1.5","border","border-gray-300","rounded","resize-none",3,"ngModelChange","ngModel"],[1,"flex","items-center","gap-3"],[1,"text-xs","font-medium","text-gray-700","whitespace-nowrap"],["type","range","min","1","max","10","step","1",1,"flex-1",3,"ngModelChange","ngModel"],[1,"flex","gap-3"],[1,"w-full","px-3","py-1.5","border","border-gray-300","rounded",3,"ngModelChange","ngModel"],["value","top-left"],["value","top"],["value","top-right"],["value","bottom-left"],["value","bottom"],["value","bottom-right"],[1,"flex-1","flex","items-center","gap-2"]],template:function(n,o){n&1&&(s(0,"div",0)(1,"button",1),g("click",function(){return o.sidebarOpen=!o.sidebarOpen}),we(),s(2,"svg",2),m(3,"path",3),r()(),Te(),s(4,"aside",4)(5,"div")(6,"h2",5),a(7,"Basics"),r(),s(8,"ul",6)(9,"li")(10,"a",7),a(11,"Installation"),r()(),s(12,"li")(13,"a",8),a(14,"Add component"),r()(),s(15,"li")(16,"a",9),a(17,"Use service"),r()()()(),s(18,"div")(19,"h2",5),a(20,"Toasts types"),r(),s(21,"ul",6)(22,"li")(23,"a",10),a(24,"Basic Toast"),r()(),s(25,"li")(26,"a",11),a(27,"Info/Success/Warn/Error Toast"),r()(),s(28,"li")(29,"a",12),a(30,"HTML Message"),r()(),s(31,"li")(32,"a",13),a(33,"Custom CSS"),r()(),s(34,"li")(35,"a",14),a(36,"Long Message"),r()(),s(37,"li")(38,"a",15),a(39,"Loading Icon"),r()(),s(40,"li")(41,"a",16),a(42,"Promise Toast"),r()(),s(43,"li")(44,"a",17),a(45,"Emoji Toast"),r()()()(),s(46,"div")(47,"h2",5),a(48,"Toast config"),r(),s(49,"ul",6)(50,"li")(51,"a",18),a(52,"Close on swipe"),r()(),s(53,"li")(54,"a",19),a(55,"Beep sound"),r()(),s(56,"li")(57,"a",20),a(58,"Progress bar"),r()(),s(59,"li")(60,"a",21),a(61,"Sticky"),r()()()(),s(62,"div")(63,"h2",5),a(64,"Toasty Container"),r(),s(65,"ul",6)(66,"li")(67,"a",22),a(68,"Default settings"),r()()()(),s(69,"div")(70,"h2",5),a(71,"API"),r(),s(72,"ul",6)(73,"li")(74,"a",23),a(75,"showToast()"),r()(),s(76,"li")(77,"a",23),a(78,"showToastPromise()"),r()()()()(),s(79,"div",24)(80,"header",25)(81,"h1",26),a(82),r(),s(83,"p",27),a(84,"Lightweight and flexible toast notifications for Angular"),r(),s(85,"div",28)(86,"button",29),g("click",function(){return o.randomToast()}),a(87," Make a Toast "),r(),s(88,"a",30),a(89," GitHub \u2197 "),r()()(),s(90,"section",31)(91,"h2",32),a(92,"Installation"),r(),s(93,"label",33),a(94,"Install via npm"),r(),m(95,"app-code",34),r(),s(96,"section",35)(97,"h2",32),a(98,"Add component"),r(),s(99,"label",36),a(100,"Add this placeholder component with the following code in your html and check out you have the required imports in your component.ts"),r(),m(101,"app-code",37),s(102,"div",38),a(103,"The container can be customized for default settings: position, capacity and duration. See "),s(104,"a",39),a(105,"Container settings"),r(),a(106," to get detailed information."),r(),s(107,"div",40),m(108,"app-code",41),r()(),s(109,"section",42)(110,"h2",32),a(111,"Service"),r(),s(112,"label",43),a(113,"Add toasty service to your component and use it to display toasts."),r(),m(114,"app-code",44),s(115,"div")(116,"label",43),a(117,"Display a basic toast!"),r(),m(118,"app-code",45),r()(),s(119,"section",46)(120,"h2",32),a(121,"Basic Toast"),r(),s(122,"label",43),a(123,"This is the most elemental toast."),r(),m(124,"app-code",47),r(),s(125,"section",48)(126,"h2",32),a(127,"Info/Success/Warning/Error Toast"),r(),s(128,"label",43),a(129,"Some customized toasts with color and icon."),r(),s(130,"button",49),g("click",function(){return o.infoToast()}),a(131,"Info"),r(),s(132,"button",49),g("click",function(){return o.successToast()}),a(133,"Sucess"),r(),s(134,"button",49),g("click",function(){return o.WarningToast()}),a(135,"Warning"),r(),s(136,"button",49),g("click",function(){return o.ErrorToast()}),a(137,"Error"),r(),m(138,"app-code",50),r(),s(139,"section",51)(140,"h2",32),a(141,"HTML Message"),r(),s(142,"p",52),a(143,"You can use HTML content in your toast message if you indicate enableHtml: true as option."),r(),m(144,"app-code",53)(145,"app-code",53)(146,"app-code",53),r(),s(147,"section",54)(148,"h2",32),a(149,"Custom CSS Toast"),r(),a(150," It is possible to create custom css rules indicating type: Custom and customStyle parameters. "),m(151,"app-code",55)(152,"app-code",56)(153,"app-code",57),a(154," Note that you can define key:value pairs just as you would in a CSS document inside a variable and assign them to "),s(155,"b"),a(156,"customStyle: style"),r()(),s(157,"section",58)(158,"h2",32),a(159,"Long Message Toast"),r(),a(160," Toast is adapted vertically in base of the content length. "),m(161,"app-code",59),r(),s(162,"section",60)(163,"h2",32),a(164,"Loading Icon Toast"),r(),a(165," Aditionally you can add a loding icon, just setting loading: true in the configuration parameter. "),m(166,"app-code",61),r(),s(167,"section",62)(168,"h2",32),a(169,"Promise Toast"),r(),s(170,"div",63),a(171,"It is possible to associate a promise. For this task you should use showToastPromise method that provides two arguments; a title and a ToastyPromise model. You can see in the following code an example."),r(),m(172,"app-code",64),s(173,"button",49),g("click",function(){return o.clickPromiseToast()}),a(174,"Promise toast example"),r()(),s(175,"section",65)(176,"h2",32),a(177,"Emoji Toast"),r(),s(178,"div",63),a(179,"Add emoji in the toasty simple using it directly in the message body:"),r(),m(180,"app-code",66),r(),s(181,"section",67)(182,"h2",32),a(183,"Close on swipe"),r(),s(184,"div",63),a(185,"You can close any toast by swiping it horizontally. This is useful for mobile devices."),r(),m(186,"app-code",68),r(),s(187,"section",69)(188,"h2",32),a(189,"Beep sound"),r(),s(190,"div",63),a(191,"You can enable a beep sound when a toast is displayed. This is useful for notifications that require attention. By default beeps are false"),r(),m(192,"app-code",70),r(),s(193,"section",71)(194,"h2",32),a(195,"Progress Bar"),r(),s(196,"div",63),a(197,"You can add a progress bar to your toast. It will animate while the toast is visible."),r(),m(198,"app-code",72),r(),s(199,"section",73)(200,"h2",32),a(201,"Sticky Toast"),r(),s(202,"div",63),a(203,"You can create a sticky toast that will not disappear automatically. The user must close it manually."),r(),m(204,"app-code",74),r(),s(205,"section",75)(206,"h2",32),a(207,"Container settings"),r(),s(208,"div",63),a(209,"You can set some default settinngs for toasts container."),r(),m(210,"app-code",76)(211,"app-code",77)(212,"app-code",77)(213,"app-code",78),s(214,"section",79)(215,"span",80),a(216,"Component parameters"),r(),s(217,"div",81)(218,"table",82)(219,"thead",83)(220,"tr")(221,"th",84),a(222,"Name"),r(),s(223,"th",84),a(224,"Type"),r(),s(225,"th",84),a(226,"Default"),r(),s(227,"th",84),a(228,"Description"),r()()(),s(229,"tbody",85)(230,"tr")(231,"td",86),a(232,"position"),r(),s(233,"td",87),a(234,"string"),r(),s(235,"td",88),a(236,'"bottom-right"'),r(),s(237,"td",87),a(238,"Controls the toast position. Options: "),s(239,"code"),a(240,"top-left"),r(),a(241,", "),s(242,"code"),a(243,"top"),r(),a(244,", "),s(245,"code"),a(246,"top-right"),r(),a(247,", "),s(248,"code"),a(249,"bottom-left"),r(),a(250,", "),s(251,"code"),a(252,"bottom"),r(),a(253,", "),s(254,"code"),a(255,"bottom-right"),r(),a(256,". "),r()(),s(257,"tr")(258,"td",86),a(259,"duration"),r(),s(260,"td",87),a(261,"number"),r(),s(262,"td",88),a(263,"5"),r(),s(264,"td",87),a(265,"Time in seconds before a toast disappears automatically."),r()(),s(266,"tr")(267,"td",86),a(268,"capacity"),r(),s(269,"td",87),a(270,"number"),r(),s(271,"td",88),a(272,"10"),r(),s(273,"td",87),a(274,"Maximum number of concurrent toasts allowed."),r()()()()()()(),s(275,"section",89)(276,"h2",32),a(277,"Demo"),r(),s(278,"div",90)(279,"h3",91),a(280,"Toasty settings"),r(),s(281,"div",92)(282,"div",93)(283,"label",94),a(284,"Title"),r(),s(285,"input",95),I("ngModelChange",function(u){return F(o.title,u)||(o.title=u),u}),r()(),s(286,"div",93)(287,"label",94),a(288,"Message"),r(),s(289,"textarea",96),I("ngModelChange",function(u){return F(o.message,u)||(o.message=u),u}),r()()(),s(290,"div",97)(291,"label",98),a(292),r(),s(293,"input",99),I("ngModelChange",function(u){return F(o.durationToast,u)||(o.durationToast=u),u}),r()()(),s(294,"div",90)(295,"h3",91),a(296,"Toasty container settings"),r(),s(297,"div",100)(298,"div",93)(299,"label",94),a(300,"Position"),r(),s(301,"select",101),I("ngModelChange",function(u){return F(o.containerPosition,u)||(o.containerPosition=u),u}),s(302,"option",102),a(303,"Top left"),r(),s(304,"option",103),a(305,"Top"),r(),s(306,"option",104),a(307,"Top right"),r(),s(308,"option",105),a(309,"Bottom left"),r(),s(310,"option",106),a(311,"Bottom"),r(),s(312,"option",107),a(313,"Bottom right"),r()()(),s(314,"div",108)(315,"label",98),a(316),r(),s(317,"input",99),I("ngModelChange",function(u){return F(o.containerCapacity,u)||(o.containerCapacity=u),u}),r()()()()(),m(318,"toasty"),r()()),n&2&&(d(4),w("-translate-x-full",!o.sidebarOpen)("translate-x-0",o.sidebarOpen)("w-256",o.sidebarOpen),d(78),C(o.libName),d(13),c("innerCode","npm install "+o.libName)("run",!1),d(6),c("innerCode","<toasty></toasty>")("run",!1),d(7),c("innerCode",o.componentcode)("run",!1),d(6),c("innerCode","constructor(private _toastyService: ToastyService) { }")("run",!1),d(4),c("innerCode","this._toastyService.showToast('Title', 'Message body');")("run",!0),d(6),c("innerCode","this._toastyService.showToast('Title', 'Message body');")("run",!0),d(14),c("innerCode",o.infosuccessetctoasts),d(6),c("title","Basic toast with html list")("innerCode","this._toastyService.showToast('Html', '<b>Details list</b><ul><li>item 1</li><li>item 2</li></ul>', { type: ToastType.Basic, enableHtml: true });"),d(),c("title","Success toast with html content & gif")("innerCode","this._toastyService.showToast('Html', 'Data successfully saved. <img width=80 src=./assets/img/dancing-spiderman.gif />', { type: ToastType.Success, enableHtml: true });"),d(),c("title","Info toast example with achievement unlocked")("innerCode","this._toastyService.showToast('Wizard', 'Achievement unlocked. <img width=100 src=./assets/img/wizard.gif />', { type: ToastType.Info, enableHtml: true });"),d(5),c("innerCode",o.customcode),d(),c("innerCode",o.customcode2),d(),c("innerCode",o.customcode3),d(8),c("innerCode",o.longToast),d(5),c("innerCode",o.loadingToast),d(6),c("innerCode",o.promiseToast)("run",!1),d(8),c("innerCode","this._toastyService.showToast('Error', '\u2665\uFE0Flove emoji toast\u2665\uFE0F');"),d(6),c("innerCode","this._toastyService.showToast('Swipe', 'Swipe me to close');")("run",!0),d(6),c("innerCode","this._toastyService.showToast('Beep', 'This toast will beep',{ beep: true });")("run",!0),d(6),c("innerCode","this._toastyService.showToast('Progress', 'This toast has a progress bar', { progressBar: true });")("run",!0),d(6),c("innerCode","this._toastyService.showToast('Sticky', 'This toast is sticky and will not disappear automatically.', { sticky: true });")("run",!0),d(6),c("innerCode",'<toasty [position]="containerPosition"></toasty>')("run",!1),d(),c("innerCode",'<toasty [duration]="containerDuration"></toasty>')("run",!1),d(),c("innerCode",'<toasty [capacity]="containerCapacity"></toasty>')("run",!1),d(),c("innerCode",'<toasty [position]="containerPosition" [duration]="containerDuration" [capacity]="containerCapacity"></toasty>')("run",!1),d(72),A("ngModel",o.title),d(4),A("ngModel",o.message),d(3),ce("Duration: ",o.durationToast,"s"),d(),A("ngModel",o.durationToast),d(8),A("ngModel",o.containerPosition),d(15),ce("Capacity: ",o.containerCapacity,""),d(),A("ngModel",o.containerCapacity))},dependencies:[Ue,pt,ct,mt,ne,be,re,at,Ce,We]})}}return t})();export{Ri as NgToastNotifyComponent};
