+++ docs/chunk-GKGHKTOS.js
@@ -1,0 +1,66 @@
+import{$a as S,Aa as U,Cb as Ge,Db as He,Ea as T,Eb as Re,Ja as V,Ka as ue,La as h,M as D,Ma as de,N as j,Na as x,O as le,Oa as b,Pa as Me,Q as M,Qa as Ae,Ra as s,S as Ve,Sa as r,Ta as p,Ua as W,Va as d,Wa as v,X as Q,Xa as Fe,Y as A,Ya as Ie,Z as ce,Za as ke,_ as C,_a as a,a as _,ab as q,b as E,bb as F,cb as I,da as K,db as k,ea as G,eb as N,fa as H,fb as O,ga as xe,gb as Ne,ha as we,hb as Oe,i as J,ia as R,jb as Pe,k as Se,ma as L,na as ee,ob as Be,p as Ee,qb as je,sb as he,ta as De,v as Te,wa as u,xa as f}from"./chunk-RBV3YYQO.js";var Le=(()=>{class t{constructor(){this.oscillatorType="sine",this.frequency=430,this.audioCtx=new AudioContext}notify(){this.audioCtx.state==="suspended"&&this.audioCtx.resume();let e=this.audioCtx.createOscillator(),n=this.audioCtx.createGain();e.type=this.oscillatorType,e.frequency.value=this.frequency,n.gain.setValueAtTime(.3,this.audioCtx.currentTime),n.gain.exponentialRampToValueAtTime(9e-5,this.audioCtx.currentTime+.4),e.connect(n),n.connect(this.audioCtx.destination),e.start(0),e.stop(this.audioCtx.currentTime+1)}static{this.\u0275fac=function(n){return new(n||t)}}static{this.\u0275prov=j({token:t,factory:t.\u0275fac,providedIn:"root"})}}return t})();var me=class{constructor(i=1/0){this.capacity=i,this.storage=[]}setCapacity(i){this.capacity=i}getCapacity(){return this.capacity}enqueue(i){if(this.size()>=this.capacity)throw Error("Toasty queue has reached max capacity, you cannot add more items.");this.storage.push(i)}dequeue(){return this.storage.shift()}size(){return this.storage.length}getElements(){return this.storage}removeId(i){let e=this.storage.findIndex(n=>n.id===i);if(e!==-1){let[n]=this.storage.splice(e,1);return n}}},m=function(t){return t.Basic="basic",t.Success="success",t.Error="error",t.Info="info",t.Warning="warning",t.Custom="custom",t}(m||{}),P=(()=>{class t{constructor(e){this._toastySoundService=e,this.TOASTY_SERVICE_CONFIG={capacity:10,duration:5e3,grouping:!1,messages:{now:"just now",seconds:"{0} seconds ago",minutes:"{0} minutes ago"},threshold_offsetX:170},this.queue=new me(this.TOASTY_SERVICE_CONFIG.capacity),this.counter=0,this.newToastBehaviorSubject=new J([]),this.newToast$=this.newToastBehaviorSubject.asObservable(),this.deleteToastBehaviorSubject=new J(void 0),this.deleteToast$=this.deleteToastBehaviorSubject.asObservable(),this.updateToastBehaviorSubject=new J(void 0),this.updateToast$=this.updateToastBehaviorSubject.asObservable()}setDefaultDuration(e){this.TOASTY_SERVICE_CONFIG.duration=e}getDefaultDuration(){return this.TOASTY_SERVICE_CONFIG.duration}setCapacity(e){this.TOASTY_SERVICE_CONFIG.capacity=e,this.queue.setCapacity(e)}getCapacity(){return this.queue.getCapacity()}getGrouping(){return this.TOASTY_SERVICE_CONFIG.grouping}setGrouping(e){this.TOASTY_SERVICE_CONFIG.grouping=e}showToast(e,n,o,l){let c=this.hash(e+n+(o?JSON.stringify(o):""));if(this.TOASTY_SERVICE_CONFIG.grouping){let w=this.queue.getElements().find(ft=>ft.hash===c);if(w)return w.count++,w.id}this.counter++;let g=o?.sticky||l?1/0:o?.duration??this.TOASTY_SERVICE_CONFIG.duration,y={id:this.counter,hash:c,title:e,message:n,type:o?.type??m.Basic,timestamp:Date.now(),duration:g,expires:Date.now()+g,config:o,count:1,get icon(){return o?.type==m.Success?"\u2705":o?.type==m.Info?"\u2139\uFE0F":o?.type==m.Warning?"\u26A0\uFE0F":o?.type==m.Error?"\u274C":""}};return this.queue.enqueue(y),this.newToastBehaviorSubject.next([...this.queue.getElements()]),y.config?.beep&&this._toastySoundService.notify(),y.config?.sticky||(y.config?.progressBar&&setTimeout(()=>{this.updateToast(y.id,{animatePbar:!0})},100),l||setTimeout(()=>{this.removeToast(y.id)},g+5)),this.counter}showToastPromise(e,n,o){let l=this.showToast(e,n.loading,E(_({},o),{loading:!0}),!0),c=o?.duration??this.TOASTY_SERVICE_CONFIG.duration;n.promise.then(g=>{let y=typeof n.success=="function"?n.success(g):n.success;this.updateToast(l,{message:y,type:m.Success,expires:Date.now()+c,config:{loading:!1}})}).catch(g=>{let y=typeof n.error=="function"?n.error(g):n.error;this.updateToast(l,{message:y,type:m.Error,expires:Date.now()+c,config:{loading:!1}})}).finally(()=>{o?.sticky||setTimeout(()=>this.removeToast(l),c)})}removeToast(e){this.newToastBehaviorSubject.next([...this.queue.getElements()]),setTimeout(()=>{let n=this.queue.removeId(e);this.newToastBehaviorSubject.next([...this.queue.getElements()])},500)}closeToast(e){let n=this.queue.removeId(e);this.newToastBehaviorSubject.next([...this.queue.getElements()])}updateTouchStart(e,n){let o={date:Date.now(),touchStart:n};this.updateToast(e,{gesture:o})}updateTouchEnd(e){let n=this.queue.getElements().forEach(o=>{if(o.gesture){let l=Math.abs(e.clientY-o.gesture.touchStart.clientY),c=Math.abs(e.clientX-o.gesture.touchStart.clientX);Date.now()-o.gesture.date<500&&c>this.TOASTY_SERVICE_CONFIG.threshold_offsetX?this.closeToast(o.id):o.gesture=void 0}})}updateTouchMove(e){this.queue.getElements().forEach(o=>{if(o.gesture){let l=e.clientX-o.gesture.touchStart.clientX;o.gesture.offsetX=l}})}updateToast(e,n){let o=this.queue.getElements(),l=o.findIndex(c=>c.id===e);l!==-1&&(o[l]=_(_({},o[l]),n),this.newToastBehaviorSubject.next([...o]))}hash(e){let n=5381;for(let o=0;o<e.length;o++)n=n*33^e.charCodeAt(o);return(n>>>0).toString(16)}static{this.\u0275fac=function(n){return new(n||t)(Ve(Le))}}static{this.\u0275prov=j({token:t,factory:t.\u0275fac,providedIn:"root"})}}return t})();var _t=(t,i)=>i.id;function vt(t,i){if(t&1&&(s(0,"div",4),a(1),r()),t&2){let e=v().$implicit;u(),S(e.icon)}}function Ct(t,i){t&1&&p(0,"div",11)}function bt(t,i){t&1&&p(0,"div",12)}function St(t,i){if(t&1&&(s(0,"span",14),a(1),r()),t&2){let e=v(2).$implicit;u(),q("x",e.count,"")}}function Et(t,i){if(t&1&&(s(0,"div",6)(1,"strong",13),a(2),r(),V(3,St,2,1,"span",14),r()),t&2){let e=v().$implicit;u(2),S(e.title),u(),b(3,e.count>1?3:-1)}}function Tt(t,i){if(t&1&&p(0,"div",7),t&2){let e=v().$implicit;h("innerHTML",e.message,De)}}function Vt(t,i){if(t&1&&(s(0,"div",15),a(1),r()),t&2){let e=v().$implicit;u(),S(e.message)}}function xt(t,i){if(t&1&&p(0,"div",16),t&2){let e=v().$implicit;de("transition-duration",e.duration,"ms"),x("animate",e.animatePbar)}}function wt(t,i){if(t&1){let e=W();s(0,"div",2)(1,"div",3),V(2,vt,2,1,"div",4)(3,Ct,1,0)(4,bt,1,0),s(5,"div",5),V(6,Et,4,2,"div",6)(7,Tt,1,1,"div",7)(8,Vt,2,1),r(),s(9,"div",8)(10,"button",9),d("click",function(){let o=G(e).$implicit,l=v();return H(l.closeToast(o.id))}),a(11,"\xD7"),r()(),V(12,xt,1,4,"div",10),r()()}if(t&2){let e=i.$implicit,n=v();de("transform","translateX("+((e.gesture==null?null:e.gesture.offsetX)||0)+"px)")("transition",e.gesture?"none":"transform 0.3s ease"),u(),x("fadeout",n.expires(e)),h("ngClass",e.type)("ngStyle",e.config==null?null:e.config.customStyle),ue("data-toast-id",e.id),u(),b(2,e.icon?2:e.config!=null&&e.config.loading?3:4),u(4),b(6,e.title?6:-1),u(),b(7,e.config!=null&&e.config.enableHtml?7:8),u(5),b(12,e.config!=null&&e.config.progressBar?12:-1)}}var pe=function(t){return t.TOP_LEFT="top-left",t.TOP="top",t.TOP_RIGHT="top-right",t.BOTTOM_LEFT="bottom-left",t.BOTTOM="bottom",t.BOTTOM_RIGHT="bottom-right",t}(pe||{}),Ue=(()=>{class t{constructor(e){this._toastService=e,this.position=pe.BOTTOM_RIGHT,this.duration=this._toastService.getDefaultDuration(),this.capacity=this._toastService.getCapacity(),this.grouping=this._toastService.getGrouping(),this.toasts$=this._toastService.newToast$,this.positionMap={"top-left":["top","left"],top:["top","center"],"top-right":["top","right"],"bottom-left":["bottom","left"],bottom:["bottom","center"],"bottom-right":["bottom","right"]}}ngOnInit(){this._toastService.deleteToast$.subscribe(e=>{}),this._toastService.updateToast$.subscribe(e=>{})}ngOnChanges(e){e.duration&&this._toastService.setDefaultDuration(this.duration),e.capacity&&this._toastService.setCapacity(e.capacity.currentValue),e.grouping&&this._toastService.setGrouping(e.grouping.currentValue)}closeToast(e){this._toastService.closeToast(e)}containerClass(){return this.positionMap[this.position]??[pe.BOTTOM_RIGHT]}expires(e){return e.expires<Date.now()}getDurationUntilExpire(e){let n=e-Date.now();return n>0?n:0}onTouchStart(e){let n=this.localizeToastIdFromTouch(e);this._toastService.updateTouchStart(n,e.changedTouches[0])}onTouchEnd(e){this._toastService.updateTouchEnd(e.changedTouches[0])}onTouchMove(e){this._toastService.updateTouchMove(e.changedTouches[0])}localizeToastIdFromTouch(e){let n=e.target,o=4;for(;!n.classList.contains("toasty")&&o>0;)n=n.parentElement,o--;return parseInt(n.getAttribute("data-toast-id")||"0")}static{this.\u0275fac=function(n){return new(n||t)(f(P))}}static{this.\u0275cmp=A({type:t,selectors:[["toasty"]],hostBindings:function(n,o){n&1&&d("touchstart",function(c){return o.onTouchStart(c)})("touchend",function(c){return o.onTouchEnd(c)})("touchmove",function(c){return o.onTouchMove(c)})},inputs:{position:"position",duration:"duration",capacity:"capacity",grouping:"grouping"},standalone:!0,features:[K,O],decls:4,vars:3,consts:[[1,"toasty-container",3,"ngClass"],[1,"toasty-wrapper",3,"transform","transition"],[1,"toasty-wrapper"],[1,"toasty",3,"ngClass","ngStyle"],[1,"toasty-icon"],[1,"toasty-content"],[1,"toasty-header"],[1,"toasty-body",3,"innerHTML"],[1,"toasty-close"],["type","button",1,"toasty-close",3,"click"],[1,"toasty-progress",3,"animate","transition-duration"],[1,"toasty-loading"],[1,"toasty-noicon"],[1,"toasty-title"],[1,"toasty-badge"],[1,"toasty-body"],[1,"toasty-progress"]],template:function(n,o){n&1&&(s(0,"div",0),Me(1,wt,13,13,"div",1,_t),Ne(3,"async"),r()),n&2&&(h("ngClass",o.containerClass()),u(),Ae(Oe(3,1,o.toasts$)))},dependencies:[Ge,He,Re],styles:[".toasty-container[_ngcontent-%COMP%]{position:fixed;width:350px;padding:.5rem;right:0;bottom:0;pointer-events:none}.toasty-container.left[_ngcontent-%COMP%]{left:12px}.toasty-container.center[_ngcontent-%COMP%]{left:50%;transform:translate(-50%)}.toasty-container.right[_ngcontent-%COMP%]{right:12px}.toasty-container.top[_ngcontent-%COMP%]{top:10px;display:flex;flex-direction:column-reverse;justify-content:flex-end}.toasty-container.bottom[_ngcontent-%COMP%]{bottom:0}.toasty[_ngcontent-%COMP%]{padding:10px;border:1px solid rgba(0,0,0,.15);border-radius:8px;box-shadow:0 4px 10px #00000026;background-color:#fff;overflow:hidden;margin-bottom:1rem;font-family:Arial,Helvetica,sans-serif;animation-name:_ngcontent-%COMP%_toasty_fadein;animation-duration:.3s;animation-timing-function:ease-out;animation-delay:0s;animation-iteration-count:1;animation-fill-mode:forwards;display:flex;align-items:center;pointer-events:auto;touch-action:manipulation}@keyframes _ngcontent-%COMP%_toasty_fadein{0%{opacity:0;transform:translate(100%)}to{opacity:1;transform:translate(0)}}@keyframes _ngcontent-%COMP%_toasty_fadeout{0%{opacity:1}to{opacity:0;pointer-events:none}}.toasty.fadeout[_ngcontent-%COMP%]{animation-name:_ngcontent-%COMP%_toasty_fadeout;animation-duration:.5s;animation-timing-function:ease-in;animation-delay:0s;animation-iteration-count:1;animation-fill-mode:forwards;pointer-events:none}.toasty-icon[_ngcontent-%COMP%]{width:20px;height:20px;margin-right:.5rem;border-radius:4px}.toasty-loading[_ngcontent-%COMP%]{width:10px;height:10px;margin-right:.6rem;border-radius:50%;display:inline-block;border-top:3px solid;border-right:3px solid transparent;box-sizing:border-box;animation:_ngcontent-%COMP%_loading-toasty-anim 1s linear infinite}@keyframes _ngcontent-%COMP%_loading-toasty-anim{0%{transform:rotate(0)}to{transform:rotate(360deg)}}.toasty-noicon[_ngcontent-%COMP%]{width:7px}.toasty-content[_ngcontent-%COMP%]{flex-direction:column;flex:1}.toasty-header[_ngcontent-%COMP%]{display:flex;align-items:center}.toasty-title[_ngcontent-%COMP%]{font-size:.9rem;font-weight:700;margin-right:auto}.toasty-time[_ngcontent-%COMP%]{font-size:.85rem;color:#6c757d;margin-right:1rem}.toasty-close[_ngcontent-%COMP%]{background:none;border:none;font-size:1.4rem;line-height:1;cursor:pointer;color:inherit;opacity:.6}.toasty-body[_ngcontent-%COMP%]{font-size:.85rem;margin-top:1px}.toasty-close[_ngcontent-%COMP%]:hover{opacity:1}.toasty.none[_ngcontent-%COMP%]{background-color:#fff;border-color:#ededed;color:#171717}.toasty.success[_ngcontent-%COMP%]{background-color:#ecfdf3;border-color:#bffcd9;color:#008a2e}.toasty.error[_ngcontent-%COMP%]{background-color:#fff0f0;border-color:#ffe0e1;color:#e60000}.toasty.info[_ngcontent-%COMP%]{background-color:#f0f8ff;border-color:#dde7fd;color:#0973dc}.toasty.warning[_ngcontent-%COMP%]{background-color:#fffcf0;border-color:#fbeeb1;color:#dc7609}.toasty.custom[_ngcontent-%COMP%]{background-color:#fff}.toasty-progress[_ngcontent-%COMP%]{position:absolute;bottom:0;left:0;height:4px;width:100%;background-color:currentColor;opacity:.3;transform-origin:left;transform:scaleX(1);transition-property:transform;transition-timing-function:linear}.toasty-progress.animate[_ngcontent-%COMP%]{transform:scaleX(0)}.toasty-badge[_ngcontent-%COMP%]{position:absolute;top:2px;right:8px;font-size:.65rem;padding:1px 5px;border-radius:999px;font-weight:700;background:#0000001a;color:currentColor}"]})}}return t})();var Mt=["codeElement"];function At(t,i){if(t&1&&(s(0,"div",2)(1,"h4",7),a(2),r()()),t&2){let e=v();u(2),S(e.title)}}function Ft(t,i){if(t&1){let e=W();s(0,"button",8),d("click",function(){G(e);let o=v();return H(o.executeCode())}),a(1," Run "),r()}}var We=(()=>{class t{constructor(e){this._toastyService=e,this.title="",this.innerCode="",this.run=!0}ngAfterViewInit(){this.highlightCode()}highlightCode(){this.codeElement&&typeof Prism<"u"&&Prism.highlightElement(this.codeElement.nativeElement)}copy(){this._toastyService.showToast("Copied","",{type:m.Info})}executeCode(){try{new Function("_toastyService","ToastType",`
+        ${this.innerCode.replace(/this\._toastyService/g,"_toastyService")};
+      `)(this._toastyService,m)}catch(e){console.error("Error executing code:",e)}}static{this.\u0275fac=function(n){return new(n||t)(f(P))}}static{this.\u0275cmp=A({type:t,selectors:[["app-code"]],viewQuery:function(n,o){if(n&1&&Fe(Mt,5),n&2){let l;Ie(l=ke())&&(o.codeElement=l.first)}},inputs:{title:"title",innerCode:"innerCode",run:"run"},standalone:!0,features:[O],decls:10,vars:3,consts:[["codeElement",""],[1,"codediv","relative"],[1,"bg-gray-200","px-4","py-2","border-b","border-gray-300","rounded-t"],[1,"bg-gray-100","text-gray-800","p-3","rounded","border","border-gray-300","mb-2"],[1,"text-sm","language-typescript"],["onclick","navigator.clipboard.writeText(this.parentElement.querySelector('code').innerText)",1,"cursor-pointer","absolute","top-2","right-2","px-2","py-1","text-xs","bg-gray-200","hover:bg-gray-300","rounded",3,"click"],[1,"absolute","bottom-2","right-2","px-2","py-1","text-xs","bg-blue-200","hover:bg-blue-300","rounded"],[1,"text-sm","font-semibold","text-gray-700","m-0"],[1,"absolute","bottom-2","right-2","px-2","py-1","text-xs","bg-blue-200","hover:bg-blue-300","rounded",3,"click"]],template:function(n,o){if(n&1){let l=W();s(0,"div",1),V(1,At,3,1,"div",2),s(2,"div",3)(3,"pre")(4,"code",4,0),a(6),r()(),s(7,"button",5),d("click",function(){return G(l),H(o.copy())}),a(8," Copy "),r(),V(9,Ft,2,0,"button",6),r()()}n&2&&(u(),b(1,o.title?1:-1),u(5),S(o.innerCode),u(3),b(9,o.run?9:-1))},styles:[".codediv[_ngcontent-%COMP%]{width:100%;overflow:auto;margin-top:4px}code[_ngcontent-%COMP%]{white-space:break-spaces}pre[_ngcontent-%COMP%]{padding:0;margin:0}"]})}}return t})();var Ke=(()=>{class t{constructor(e,n){this._renderer=e,this._elementRef=n,this.onChange=o=>{},this.onTouched=()=>{}}setProperty(e,n){this._renderer.setProperty(this._elementRef.nativeElement,e,n)}registerOnTouched(e){this.onTouched=e}registerOnChange(e){this.onChange=e}setDisabledState(e){this.setProperty("disabled",e)}static{this.\u0275fac=function(n){return new(n||t)(f(U),f(L))}}static{this.\u0275dir=C({type:t})}}return t})(),oe=(()=>{class t extends Ke{static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=R(t)))(o||t)}})()}static{this.\u0275dir=C({type:t,features:[T]})}}return t})(),Z=new M("");var It={provide:Z,useExisting:D(()=>re),multi:!0};function kt(){let t=he()?he().getUserAgent():"";return/android (\d+)/.test(t.toLowerCase())}var Nt=new M(""),re=(()=>{class t extends Ke{constructor(e,n,o){super(e,n),this._compositionMode=o,this._composing=!1,this._compositionMode==null&&(this._compositionMode=!kt())}writeValue(e){let n=e??"";this.setProperty("value",n)}_handleInput(e){(!this._compositionMode||this._compositionMode&&!this._composing)&&this.onChange(e)}_compositionStart(){this._composing=!0}_compositionEnd(e){this._composing=!1,this._compositionMode&&this.onChange(e)}static{this.\u0275fac=function(n){return new(n||t)(f(U),f(L),f(Nt,8))}}static{this.\u0275dir=C({type:t,selectors:[["input","formControlName","",3,"type","checkbox"],["textarea","formControlName",""],["input","formControl","",3,"type","checkbox"],["textarea","formControl",""],["input","ngModel","",3,"type","checkbox"],["textarea","ngModel",""],["","ngDefaultControl",""]],hostBindings:function(n,o){n&1&&d("input",function(c){return o._handleInput(c.target.value)})("blur",function(){return o.onTouched()})("compositionstart",function(){return o._compositionStart()})("compositionend",function(c){return o._compositionEnd(c.target.value)})},features:[N([It]),T]})}}return t})();var Ot=new M(""),Pt=new M("");function et(t){return t!=null}function tt(t){return Pe(t)?Se(t):t}function it(t){let i={};return t.forEach(e=>{i=e!=null?_(_({},i),e):i}),Object.keys(i).length===0?null:i}function nt(t,i){return i.map(e=>e(t))}function Bt(t){return!t.validate}function ot(t){return t.map(i=>Bt(i)?i:e=>i.validate(e))}function jt(t){if(!t)return null;let i=t.filter(et);return i.length==0?null:function(e){return it(nt(e,i))}}function rt(t){return t!=null?jt(ot(t)):null}function Gt(t){if(!t)return null;let i=t.filter(et);return i.length==0?null:function(e){let n=nt(e,i).map(tt);return Te(n).pipe(Ee(it))}}function st(t){return t!=null?Gt(ot(t)):null}function qe(t,i){return t===null?[i]:Array.isArray(t)?[...t,i]:[t,i]}function Ht(t){return t._rawValidators}function Rt(t){return t._rawAsyncValidators}function fe(t){return t?Array.isArray(t)?t:[t]:[]}function ie(t,i){return Array.isArray(t)?t.includes(i):t===i}function $e(t,i){let e=fe(i);return fe(t).forEach(o=>{ie(e,o)||e.push(o)}),e}function ze(t,i){return fe(i).filter(e=>!ie(t,e))}var ne=class{constructor(){this._rawValidators=[],this._rawAsyncValidators=[],this._onDestroyCallbacks=[]}get value(){return this.control?this.control.value:null}get valid(){return this.control?this.control.valid:null}get invalid(){return this.control?this.control.invalid:null}get pending(){return this.control?this.control.pending:null}get disabled(){return this.control?this.control.disabled:null}get enabled(){return this.control?this.control.enabled:null}get errors(){return this.control?this.control.errors:null}get pristine(){return this.control?this.control.pristine:null}get dirty(){return this.control?this.control.dirty:null}get touched(){return this.control?this.control.touched:null}get status(){return this.control?this.control.status:null}get untouched(){return this.control?this.control.untouched:null}get statusChanges(){return this.control?this.control.statusChanges:null}get valueChanges(){return this.control?this.control.valueChanges:null}get path(){return null}_setValidators(i){this._rawValidators=i||[],this._composedValidatorFn=rt(this._rawValidators)}_setAsyncValidators(i){this._rawAsyncValidators=i||[],this._composedAsyncValidatorFn=st(this._rawAsyncValidators)}get validator(){return this._composedValidatorFn||null}get asyncValidator(){return this._composedAsyncValidatorFn||null}_registerOnDestroy(i){this._onDestroyCallbacks.push(i)}_invokeOnDestroyCallbacks(){this._onDestroyCallbacks.forEach(i=>i()),this._onDestroyCallbacks=[]}reset(i=void 0){this.control&&this.control.reset(i)}hasError(i,e){return this.control?this.control.hasError(i,e):!1}getError(i,e){return this.control?this.control.getError(i,e):null}},ge=class extends ne{get formDirective(){return null}get path(){return null}},X=class extends ne{constructor(){super(...arguments),this._parent=null,this.name=null,this.valueAccessor=null}},ye=class{constructor(i){this._cd=i}get isTouched(){return!!this._cd?.control?.touched}get isUntouched(){return!!this._cd?.control?.untouched}get isPristine(){return!!this._cd?.control?.pristine}get isDirty(){return!!this._cd?.control?.dirty}get isValid(){return!!this._cd?.control?.valid}get isInvalid(){return!!this._cd?.control?.invalid}get isPending(){return!!this._cd?.control?.pending}get isSubmitted(){return!!this._cd?.submitted}},Lt={"[class.ng-untouched]":"isUntouched","[class.ng-touched]":"isTouched","[class.ng-pristine]":"isPristine","[class.ng-dirty]":"isDirty","[class.ng-valid]":"isValid","[class.ng-invalid]":"isInvalid","[class.ng-pending]":"isPending"},Bi=E(_({},Lt),{"[class.ng-submitted]":"isSubmitted"}),at=(()=>{class t extends ye{constructor(e){super(e)}static{this.\u0275fac=function(n){return new(n||t)(f(X,2))}}static{this.\u0275dir=C({type:t,selectors:[["","formControlName",""],["","ngModel",""],["","formControl",""]],hostVars:14,hostBindings:function(n,o){n&2&&x("ng-untouched",o.isUntouched)("ng-touched",o.isTouched)("ng-pristine",o.isPristine)("ng-dirty",o.isDirty)("ng-valid",o.isValid)("ng-invalid",o.isInvalid)("ng-pending",o.isPending)},features:[T]})}}return t})();var z="VALID",te="INVALID",B="PENDING",Y="DISABLED";function Ut(t){return(se(t)?t.validators:t)||null}function Wt(t){return Array.isArray(t)?rt(t):t||null}function qt(t,i){return(se(i)?i.asyncValidators:t)||null}function $t(t){return Array.isArray(t)?st(t):t||null}function se(t){return t!=null&&!Array.isArray(t)&&typeof t=="object"}var _e=class{constructor(i,e){this._pendingDirty=!1,this._hasOwnPendingAsyncValidator=!1,this._pendingTouched=!1,this._onCollectionChange=()=>{},this._parent=null,this.pristine=!0,this.touched=!1,this._onDisabledChange=[],this._assignValidators(i),this._assignAsyncValidators(e)}get validator(){return this._composedValidatorFn}set validator(i){this._rawValidators=this._composedValidatorFn=i}get asyncValidator(){return this._composedAsyncValidatorFn}set asyncValidator(i){this._rawAsyncValidators=this._composedAsyncValidatorFn=i}get parent(){return this._parent}get valid(){return this.status===z}get invalid(){return this.status===te}get pending(){return this.status==B}get disabled(){return this.status===Y}get enabled(){return this.status!==Y}get dirty(){return!this.pristine}get untouched(){return!this.touched}get updateOn(){return this._updateOn?this._updateOn:this.parent?this.parent.updateOn:"change"}setValidators(i){this._assignValidators(i)}setAsyncValidators(i){this._assignAsyncValidators(i)}addValidators(i){this.setValidators($e(i,this._rawValidators))}addAsyncValidators(i){this.setAsyncValidators($e(i,this._rawAsyncValidators))}removeValidators(i){this.setValidators(ze(i,this._rawValidators))}removeAsyncValidators(i){this.setAsyncValidators(ze(i,this._rawAsyncValidators))}hasValidator(i){return ie(this._rawValidators,i)}hasAsyncValidator(i){return ie(this._rawAsyncValidators,i)}clearValidators(){this.validator=null}clearAsyncValidators(){this.asyncValidator=null}markAsTouched(i={}){this.touched=!0,this._parent&&!i.onlySelf&&this._parent.markAsTouched(i)}markAllAsTouched(){this.markAsTouched({onlySelf:!0}),this._forEachChild(i=>i.markAllAsTouched())}markAsUntouched(i={}){this.touched=!1,this._pendingTouched=!1,this._forEachChild(e=>{e.markAsUntouched({onlySelf:!0})}),this._parent&&!i.onlySelf&&this._parent._updateTouched(i)}markAsDirty(i={}){this.pristine=!1,this._parent&&!i.onlySelf&&this._parent.markAsDirty(i)}markAsPristine(i={}){this.pristine=!0,this._pendingDirty=!1,this._forEachChild(e=>{e.markAsPristine({onlySelf:!0})}),this._parent&&!i.onlySelf&&this._parent._updatePristine(i)}markAsPending(i={}){this.status=B,i.emitEvent!==!1&&this.statusChanges.emit(this.status),this._parent&&!i.onlySelf&&this._parent.markAsPending(i)}disable(i={}){let e=this._parentMarkedDirty(i.onlySelf);this.status=Y,this.errors=null,this._forEachChild(n=>{n.disable(E(_({},i),{onlySelf:!0}))}),this._updateValue(),i.emitEvent!==!1&&(this.valueChanges.emit(this.value),this.statusChanges.emit(this.status)),this._updateAncestors(E(_({},i),{skipPristineCheck:e})),this._onDisabledChange.forEach(n=>n(!0))}enable(i={}){let e=this._parentMarkedDirty(i.onlySelf);this.status=z,this._forEachChild(n=>{n.enable(E(_({},i),{onlySelf:!0}))}),this.updateValueAndValidity({onlySelf:!0,emitEvent:i.emitEvent}),this._updateAncestors(E(_({},i),{skipPristineCheck:e})),this._onDisabledChange.forEach(n=>n(!1))}_updateAncestors(i){this._parent&&!i.onlySelf&&(this._parent.updateValueAndValidity(i),i.skipPristineCheck||this._parent._updatePristine(),this._parent._updateTouched())}setParent(i){this._parent=i}getRawValue(){return this.value}updateValueAndValidity(i={}){this._setInitialStatus(),this._updateValue(),this.enabled&&(this._cancelExistingSubscription(),this.errors=this._runValidator(),this.status=this._calculateStatus(),(this.status===z||this.status===B)&&this._runAsyncValidator(i.emitEvent)),i.emitEvent!==!1&&(this.valueChanges.emit(this.value),this.statusChanges.emit(this.status)),this._parent&&!i.onlySelf&&this._parent.updateValueAndValidity(i)}_updateTreeValidity(i={emitEvent:!0}){this._forEachChild(e=>e._updateTreeValidity(i)),this.updateValueAndValidity({onlySelf:!0,emitEvent:i.emitEvent})}_setInitialStatus(){this.status=this._allControlsDisabled()?Y:z}_runValidator(){return this.validator?this.validator(this):null}_runAsyncValidator(i){if(this.asyncValidator){this.status=B,this._hasOwnPendingAsyncValidator=!0;let e=tt(this.asyncValidator(this));this._asyncValidationSubscription=e.subscribe(n=>{this._hasOwnPendingAsyncValidator=!1,this.setErrors(n,{emitEvent:i})})}}_cancelExistingSubscription(){this._asyncValidationSubscription&&(this._asyncValidationSubscription.unsubscribe(),this._hasOwnPendingAsyncValidator=!1)}setErrors(i,e={}){this.errors=i,this._updateControlsErrors(e.emitEvent!==!1)}get(i){let e=i;return e==null||(Array.isArray(e)||(e=e.split(".")),e.length===0)?null:e.reduce((n,o)=>n&&n._find(o),this)}getError(i,e){let n=e?this.get(e):this;return n&&n.errors?n.errors[i]:null}hasError(i,e){return!!this.getError(i,e)}get root(){let i=this;for(;i._parent;)i=i._parent;return i}_updateControlsErrors(i){this.status=this._calculateStatus(),i&&this.statusChanges.emit(this.status),this._parent&&this._parent._updateControlsErrors(i)}_initObservables(){this.valueChanges=new ee,this.statusChanges=new ee}_calculateStatus(){return this._allControlsDisabled()?Y:this.errors?te:this._hasOwnPendingAsyncValidator||this._anyControlsHaveStatus(B)?B:this._anyControlsHaveStatus(te)?te:z}_anyControlsHaveStatus(i){return this._anyControls(e=>e.status===i)}_anyControlsDirty(){return this._anyControls(i=>i.dirty)}_anyControlsTouched(){return this._anyControls(i=>i.touched)}_updatePristine(i={}){this.pristine=!this._anyControlsDirty(),this._parent&&!i.onlySelf&&this._parent._updatePristine(i)}_updateTouched(i={}){this.touched=this._anyControlsTouched(),this._parent&&!i.onlySelf&&this._parent._updateTouched(i)}_registerOnCollectionChange(i){this._onCollectionChange=i}_setUpdateStrategy(i){se(i)&&i.updateOn!=null&&(this._updateOn=i.updateOn)}_parentMarkedDirty(i){let e=this._parent&&this._parent.dirty;return!i&&!!e&&!this._parent._anyControlsDirty()}_find(i){return null}_assignValidators(i){this._rawValidators=Array.isArray(i)?i.slice():i,this._composedValidatorFn=Wt(this._rawValidators)}_assignAsyncValidators(i){this._rawAsyncValidators=Array.isArray(i)?i.slice():i,this._composedAsyncValidatorFn=$t(this._rawAsyncValidators)}};var lt=new M("CallSetDisabledState",{providedIn:"root",factory:()=>ve}),ve="always";function zt(t,i){return[...i.path,t]}function Yt(t,i,e=ve){Zt(t,i),i.valueAccessor.writeValue(t.value),(t.disabled||e==="always")&&i.valueAccessor.setDisabledState?.(t.disabled),Jt(t,i),Kt(t,i),Qt(t,i),Xt(t,i)}function Ye(t,i){t.forEach(e=>{e.registerOnValidatorChange&&e.registerOnValidatorChange(i)})}function Xt(t,i){if(i.valueAccessor.setDisabledState){let e=n=>{i.valueAccessor.setDisabledState(n)};t.registerOnDisabledChange(e),i._registerOnDestroy(()=>{t._unregisterOnDisabledChange(e)})}}function Zt(t,i){let e=Ht(t);i.validator!==null?t.setValidators(qe(e,i.validator)):typeof e=="function"&&t.setValidators([e]);let n=Rt(t);i.asyncValidator!==null?t.setAsyncValidators(qe(n,i.asyncValidator)):typeof n=="function"&&t.setAsyncValidators([n]);let o=()=>t.updateValueAndValidity();Ye(i._rawValidators,o),Ye(i._rawAsyncValidators,o)}function Jt(t,i){i.valueAccessor.registerOnChange(e=>{t._pendingValue=e,t._pendingChange=!0,t._pendingDirty=!0,t.updateOn==="change"&&ct(t,i)})}function Qt(t,i){i.valueAccessor.registerOnTouched(()=>{t._pendingTouched=!0,t.updateOn==="blur"&&t._pendingChange&&ct(t,i),t.updateOn!=="submit"&&t.markAsTouched()})}function ct(t,i){t._pendingDirty&&t.markAsDirty(),t.setValue(t._pendingValue,{emitModelToViewChange:!1}),i.viewToModelUpdate(t._pendingValue),t._pendingChange=!1}function Kt(t,i){let e=(n,o)=>{i.valueAccessor.writeValue(n),o&&i.viewToModelUpdate(n)};t.registerOnChange(e),i._registerOnDestroy(()=>{t._unregisterOnChange(e)})}function ei(t,i){if(!t.hasOwnProperty("model"))return!1;let e=t.model;return e.isFirstChange()?!0:!Object.is(i,e.currentValue)}function ti(t){return Object.getPrototypeOf(t.constructor)===oe}function ii(t,i){if(!i)return null;Array.isArray(i);let e,n,o;return i.forEach(l=>{l.constructor===re?e=l:ti(l)?n=l:o=l}),o||n||e||null}function Xe(t,i){let e=t.indexOf(i);e>-1&&t.splice(e,1)}function Ze(t){return typeof t=="object"&&t!==null&&Object.keys(t).length===2&&"value"in t&&"disabled"in t}var ni=class extends _e{constructor(i=null,e,n){super(Ut(e),qt(n,e)),this.defaultValue=null,this._onChange=[],this._pendingChange=!1,this._applyFormState(i),this._setUpdateStrategy(e),this._initObservables(),this.updateValueAndValidity({onlySelf:!0,emitEvent:!!this.asyncValidator}),se(e)&&(e.nonNullable||e.initialValueIsDefault)&&(Ze(i)?this.defaultValue=i.value:this.defaultValue=i)}setValue(i,e={}){this.value=this._pendingValue=i,this._onChange.length&&e.emitModelToViewChange!==!1&&this._onChange.forEach(n=>n(this.value,e.emitViewToModelChange!==!1)),this.updateValueAndValidity(e)}patchValue(i,e={}){this.setValue(i,e)}reset(i=this.defaultValue,e={}){this._applyFormState(i),this.markAsPristine(e),this.markAsUntouched(e),this.setValue(this.value,e),this._pendingChange=!1}_updateValue(){}_anyControls(i){return!1}_allControlsDisabled(){return this.disabled}registerOnChange(i){this._onChange.push(i)}_unregisterOnChange(i){Xe(this._onChange,i)}registerOnDisabledChange(i){this._onDisabledChange.push(i)}_unregisterOnDisabledChange(i){Xe(this._onDisabledChange,i)}_forEachChild(i){}_syncPendingControls(){return this.updateOn==="submit"&&(this._pendingDirty&&this.markAsDirty(),this._pendingTouched&&this.markAsTouched(),this._pendingChange)?(this.setValue(this._pendingValue,{onlySelf:!0,emitModelToViewChange:!1}),!0):!1}_applyFormState(i){Ze(i)?(this.value=this._pendingValue=i.value,i.disabled?this.disable({onlySelf:!0,emitEvent:!1}):this.enable({onlySelf:!0,emitEvent:!1})):this.value=this._pendingValue=i}};var oi={provide:X,useExisting:D(()=>Ce)},Je=Promise.resolve(),Ce=(()=>{class t extends X{constructor(e,n,o,l,c,g){super(),this._changeDetectorRef=c,this.callSetDisabledState=g,this.control=new ni,this._registered=!1,this.name="",this.update=new ee,this._parent=e,this._setValidators(n),this._setAsyncValidators(o),this.valueAccessor=ii(this,l)}ngOnChanges(e){if(this._checkForErrors(),!this._registered||"name"in e){if(this._registered&&(this._checkName(),this.formDirective)){let n=e.name.previousValue;this.formDirective.removeControl({name:n,path:this._getPath(n)})}this._setUpControl()}"isDisabled"in e&&this._updateDisabled(e),ei(e,this.viewModel)&&(this._updateValue(this.model),this.viewModel=this.model)}ngOnDestroy(){this.formDirective&&this.formDirective.removeControl(this)}get path(){return this._getPath(this.name)}get formDirective(){return this._parent?this._parent.formDirective:null}viewToModelUpdate(e){this.viewModel=e,this.update.emit(e)}_setUpControl(){this._setUpdateStrategy(),this._isStandalone()?this._setUpStandalone():this.formDirective.addControl(this),this._registered=!0}_setUpdateStrategy(){this.options&&this.options.updateOn!=null&&(this.control._updateOn=this.options.updateOn)}_isStandalone(){return!this._parent||!!(this.options&&this.options.standalone)}_setUpStandalone(){Yt(this.control,this,this.callSetDisabledState),this.control.updateValueAndValidity({emitEvent:!1})}_checkForErrors(){this._isStandalone()||this._checkParentType(),this._checkName()}_checkParentType(){}_checkName(){this.options&&this.options.name&&(this.name=this.options.name),!this._isStandalone()&&this.name}_updateValue(e){Je.then(()=>{this.control.setValue(e,{emitViewToModelChange:!1}),this._changeDetectorRef?.markForCheck()})}_updateDisabled(e){let n=e.isDisabled.currentValue,o=n!==0&&je(n);Je.then(()=>{o&&!this.control.disabled?this.control.disable():!o&&this.control.disabled&&this.control.enable(),this._changeDetectorRef?.markForCheck()})}_getPath(e){return this._parent?zt(e,this._parent):[e]}static{this.\u0275fac=function(n){return new(n||t)(f(ge,9),f(Ot,10),f(Pt,10),f(Z,10),f(Be,8),f(lt,8))}}static{this.\u0275dir=C({type:t,selectors:[["","ngModel","",3,"formControlName","",3,"formControl",""]],inputs:{name:"name",isDisabled:[Q.None,"disabled","isDisabled"],model:[Q.None,"ngModel","model"],options:[Q.None,"ngModelOptions","options"]},outputs:{update:"ngModelChange"},exportAs:["ngModel"],features:[N([oi]),T,K]})}}return t})();var ri={provide:Z,useExisting:D(()=>be),multi:!0},be=(()=>{class t extends oe{writeValue(e){this.setProperty("value",parseFloat(e))}registerOnChange(e){this.onChange=n=>{e(n==""?null:parseFloat(n))}}static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=R(t)))(o||t)}})()}static{this.\u0275dir=C({type:t,selectors:[["input","type","range","formControlName",""],["input","type","range","formControl",""],["input","type","range","ngModel",""]],hostBindings:function(n,o){n&1&&d("change",function(c){return o.onChange(c.target.value)})("input",function(c){return o.onChange(c.target.value)})("blur",function(){return o.onTouched()})},features:[N([ri]),T]})}}return t})();var si={provide:Z,useExisting:D(()=>ae),multi:!0};function ut(t,i){return t==null?`${i}`:(i&&typeof i=="object"&&(i="Object"),`${t}: ${i}`.slice(0,50))}function ai(t){return t.split(":")[0]}var ae=(()=>{class t extends oe{constructor(){super(...arguments),this._optionMap=new Map,this._idCounter=0,this._compareWith=Object.is}set compareWith(e){this._compareWith=e}writeValue(e){this.value=e;let n=this._getOptionId(e),o=ut(n,e);this.setProperty("value",o)}registerOnChange(e){this.onChange=n=>{this.value=this._getOptionValue(n),e(this.value)}}_registerOption(){return(this._idCounter++).toString()}_getOptionId(e){for(let n of this._optionMap.keys())if(this._compareWith(this._optionMap.get(n),e))return n;return null}_getOptionValue(e){let n=ai(e);return this._optionMap.has(n)?this._optionMap.get(n):e}static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=R(t)))(o||t)}})()}static{this.\u0275dir=C({type:t,selectors:[["select","formControlName","",3,"multiple",""],["select","formControl","",3,"multiple",""],["select","ngModel","",3,"multiple",""]],hostBindings:function(n,o){n&1&&d("change",function(c){return o.onChange(c.target.value)})("blur",function(){return o.onTouched()})},inputs:{compareWith:"compareWith"},features:[N([si]),T]})}}return t})(),dt=(()=>{class t{constructor(e,n,o){this._element=e,this._renderer=n,this._select=o,this._select&&(this.id=this._select._registerOption())}set ngValue(e){this._select!=null&&(this._select._optionMap.set(this.id,e),this._setElementValue(ut(this.id,e)),this._select.writeValue(this._select.value))}set value(e){this._setElementValue(e),this._select&&this._select.writeValue(this._select.value)}_setElementValue(e){this._renderer.setProperty(this._element.nativeElement,"value",e)}ngOnDestroy(){this._select&&(this._select._optionMap.delete(this.id),this._select.writeValue(this._select.value))}static{this.\u0275fac=function(n){return new(n||t)(f(L),f(U),f(ae,9))}}static{this.\u0275dir=C({type:t,selectors:[["option"]],inputs:{ngValue:"ngValue",value:"value"}})}}return t})(),li={provide:Z,useExisting:D(()=>ht),multi:!0};function Qe(t,i){return t==null?`${i}`:(typeof i=="string"&&(i=`'${i}'`),i&&typeof i=="object"&&(i="Object"),`${t}: ${i}`.slice(0,50))}function ci(t){return t.split(":")[0]}var ht=(()=>{class t extends oe{constructor(){super(...arguments),this._optionMap=new Map,this._idCounter=0,this._compareWith=Object.is}set compareWith(e){this._compareWith=e}writeValue(e){this.value=e;let n;if(Array.isArray(e)){let o=e.map(l=>this._getOptionId(l));n=(l,c)=>{l._setSelected(o.indexOf(c.toString())>-1)}}else n=(o,l)=>{o._setSelected(!1)};this._optionMap.forEach(n)}registerOnChange(e){this.onChange=n=>{let o=[],l=n.selectedOptions;if(l!==void 0){let c=l;for(let g=0;g<c.length;g++){let y=c[g],w=this._getOptionValue(y.value);o.push(w)}}else{let c=n.options;for(let g=0;g<c.length;g++){let y=c[g];if(y.selected){let w=this._getOptionValue(y.value);o.push(w)}}}this.value=o,e(o)}}_registerOption(e){let n=(this._idCounter++).toString();return this._optionMap.set(n,e),n}_getOptionId(e){for(let n of this._optionMap.keys())if(this._compareWith(this._optionMap.get(n)._value,e))return n;return null}_getOptionValue(e){let n=ci(e);return this._optionMap.has(n)?this._optionMap.get(n)._value:e}static{this.\u0275fac=(()=>{let e;return function(o){return(e||(e=R(t)))(o||t)}})()}static{this.\u0275dir=C({type:t,selectors:[["select","multiple","","formControlName",""],["select","multiple","","formControl",""],["select","multiple","","ngModel",""]],hostBindings:function(n,o){n&1&&d("change",function(c){return o.onChange(c.target)})("blur",function(){return o.onTouched()})},inputs:{compareWith:"compareWith"},features:[N([li]),T]})}}return t})(),mt=(()=>{class t{constructor(e,n,o){this._element=e,this._renderer=n,this._select=o,this._select&&(this.id=this._select._registerOption(this))}set ngValue(e){this._select!=null&&(this._value=e,this._setElementValue(Qe(this.id,e)),this._select.writeValue(this._select.value))}set value(e){this._select?(this._value=e,this._setElementValue(Qe(this.id,e)),this._select.writeValue(this._select.value)):this._setElementValue(e)}_setElementValue(e){this._renderer.setProperty(this._element.nativeElement,"value",e)}_setSelected(e){this._renderer.setProperty(this._element.nativeElement,"selected",e)}ngOnDestroy(){this._select&&(this._select._optionMap.delete(this.id),this._select.writeValue(this._select.value))}static{this.\u0275fac=function(n){return new(n||t)(f(L),f(U),f(ht,9))}}static{this.\u0275dir=C({type:t,selectors:[["option"]],inputs:{ngValue:"ngValue",value:"value"}})}}return t})();var ui=(()=>{class t{static{this.\u0275fac=function(n){return new(n||t)}}static{this.\u0275mod=ce({type:t})}static{this.\u0275inj=le({})}}return t})();var pt=(()=>{class t{static withConfig(e){return{ngModule:t,providers:[{provide:lt,useValue:e.callSetDisabledState??ve}]}}static{this.\u0275fac=function(n){return new(n||t)}}static{this.\u0275mod=ce({type:t})}static{this.\u0275inj=le({imports:[ui]})}}return t})();var Ui=(()=>{class t{constructor(e){this._toastyService=e,this.title="",this.message="",this.containerDuration=5e3,this.containerPosition="bottom-right",this.containerCapacity=10,this.containerGrouping=!1,this.durationToast=5,this.libName="ng-toast-notify",this.sidebarOpen=!1,this.toastMessages=[{title:"Mission Accomplished!",message:"You pressed a button. NASA is impressed."},{title:"Toast Success!",message:"This notification is brought to you by JavaScript and caffeine."},{title:"Brain.exe has stopped",message:"Just kidding! Everything\u2019s fine. Probably."},{title:"You\u2019ve Got Mail!",message:"Just kidding. It\u2019s 2025. No one sends emails anymore."},{title:"Bug? Never Heard of Her",message:"That action definitely didn\u2019t break anything. Maybe."},{title:"Toast Delivered",message:"Served fresh and slightly over-engineered."},{title:"Cool Move",message:"That click had style. Hollywood is calling."},{title:"Clean Slate",message:"We wiped it. Like, really wiped it. Hope that was okay."},{title:"Access Granted",message:"You are now officially allowed to feel powerful."},{title:"Fancy That!",message:"A toast so elegant, it sips its coffee with its pinky up."},{title:"Action Completed",message:"Your changes have been saved successfully."},{title:"Data Submitted",message:"The form was sent and processed correctly."},{title:"Warning",message:"Please check the required fields before continuing."},{title:"Error Occurred",message:"Something went wrong. Try again or contact support."},{title:"Changes Updated",message:"Your settings have been updated."},{title:"Logged In",message:"Welcome back! You have successfully signed in."},{title:"Access Denied",message:"You don\u2019t have permission to perform this action."},{title:"Session Expired",message:"Please log in again to continue."},{title:"Item Deleted",message:"The selected item has been removed from the list."},{title:"Autosaved",message:"Your progress was saved automatically."}],this.infosuccessetctoasts=`
+  this._toastyService.showToast('Info', 'Information body content', { type: ToastType.Info})
+  this._toastyService.showToast('Success', 'Success body content', { type: ToastType.Success})
+  this._toastyService.showToast('Warning', 'Warning body content', { type: ToastType.Warning})
+  this._toastyService.showToast('Error', 'Error body content', { type: ToastType.Error})
+  `,this.componentcode=`
+  import { ToastType, ToastyComponent, ToastyPromise, ToastyService } from '${this.libName}'; 
+  @Component({
+    selector: 'app-root',
+    standalone: true,
+    imports: [ToastyComponent],
+    templateUrl: './app.component.html',
+    styleUrl: './app.component.css'
+  })
+  `,this.customcode=`
+  const style = {
+      "background-image": "linear-gradient(to right, #e52d27 0%, #b31217  51%, #e52d27  100%)",
+      "color": "#fff"
+    };
+    this._toastyService.showToast('Custom', 'Some body content', { type: ToastType.Custom, customStyle: style });
+  `,this.customcode2=`
+  const style = {
+      "background-image": "linear-gradient(220.55deg, #FFE70B 0%, #27B643 100%)",
+      "border": "1px solid rgb(182 214 19)",
+      "color": "#fff"
+    };
+    this._toastyService.showToast('Custom', 'Some body content', { type: ToastType.Custom, customStyle: style });
+  `,this.customcode3=`
+    const style = { 
+      "gap": "12px",
+      "padding": "14px 28px",
+      "background": "linear-gradient(145deg, #0f0f0f, #1c1c1c)",
+      "border": "2px solid rgba(255, 255, 255, 0.2)",
+      "border-radius": "100px",
+      "color": "#fff",
+      "font-size": "16px",
+      "font-weight": "600",
+      "letter-spacing": "0.5px",
+      "cursor": "pointer",
+      "overflow": "hidden",
+      "box-shadow": "0 0 20px rgba(0, 255, 255, 0.1)",
+      "backdrop-filter": "blur(8px)"
+    };
+    this._toastyService.showToast('Custom', 'Some body content', { type: ToastType.Custom, customStyle: style });
+  `,this.longToast=`
+  this._toastyService.showToast("Long text", "Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla tincidunt euismod nunc quis suscipit. Curabitur sed semper tortor, at efficitur leo. Sed justo nulla, ullamcorper sed maximus a, vehicula ut nulla. Vivamus lobortis lacinia mi eu ultricies. Suspendisse quis sem ante. Vivamus consectetur lectus eget quam venenatis, sagittis aliquam lectus viverra. Donec nec molestie magna, ut ultricies tellus. Praesent faucibus urna imperdiet vulputate dapibus. Sed sit amet dolor odio. Donec est nulla, cursus sed aliquam at, cursus sit amet turpis. Nulla ut dolor eget dui laoreet hendrerit vel at tellus. Nullam aliquet sem at mauris scelerisque molestie. Nullam imperdiet blandit est et aliquam. Vestibulum euismod, libero id tincidunt rutrum, nulla augue aliquet risus, a rutrum elit orci eget lectus.");
+  `,this.loadingToast=`
+  this._toastyService.showToast('Wait', 'Process may take a few seconds to complete', { type: ToastType.Custom, loading:true });
+  `,this.promiseToast=`
+    simulateApiCall(): Promise<{ name: string }> {
+      return new Promise((resolve, reject) =>
+        setTimeout(
+            () => Math.random() < 0.5 ? resolve({ name: 'Toasty resolved' }) : reject(new Error('Toasty rejected'))
+        , 2000));
+    }
+
+    this._toastyService.showToastPromise('Fetching API', 
+                                          {
+                                            promise: this.simulateApiCall(),
+                                            loading: 'Loading',
+                                            success: (data) => \`Ok \${data.name}\`,
+                                            error: 'Error'
+                                          });  
+  `}basicToast(){this._toastyService.showToast("Title","Message body")}infoToast(){this._toastyService.showToast("Info","Information body content",{type:m.Info})}successToast(){this._toastyService.showToast("Success","Success body content",{type:m.Success})}WarningToast(){this._toastyService.showToast("Warning","Warning body content",{type:m.Warning})}ErrorToast(){this._toastyService.showToast("Error","Error body content",{type:m.Error})}htmlToast(){this._toastyService.showToast("Html","<b>Details</b><ul><li>item 1</li><li>item 2</li></ul>",{type:m.Basic,enableHtml:!0})}htmlToast2(){this._toastyService.showToast("Html","Data successfully saved. <img width=80 src=./assets/img/dancing-spiderman.gif /> ",{type:m.Success,enableHtml:!0})}htmlToast3(){this._toastyService.showToast("Html","Achievement unlocked. <img width=100 src=./assets/img/wizard.gif /> ",{type:m.Info,enableHtml:!0})}customStyleToast(){let e={"background-image":"linear-gradient(to right, #e52d27 0%, #b31217  51%, #e52d27  100%)",color:"#fff"};this._toastyService.showToast("Custom","Some body content",{type:m.Custom,customStyle:e})}emojiToast(){this._toastyService.showToast("Error","\u2665\uFE0Flove emoji toast\u2665\uFE0F")}makeToast(){let e=[m.Info,m.Success,m.Warning,m.Basic,m.Error],n=e[Math.floor(Math.random()*e.length)],o=Math.floor(Math.random()*this.toastMessages.length),l=[{"background-image":"linear-gradient(to right, #77A1D3 0%, #79CBCA  51%, #77A1D3  100%)",color:"#fff","font-family":"cursive"},{"background-image":"linear-gradient(to right, #e52d27 0%, #b31217  51%, #e52d27  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, #FF512F 0%, #F09819  51%, #FF512F  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, #3D7EAA 0%, #FFE47A  51%, #3D7EAA  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, #BA8B02 0%, #181818  51%, #BA8B02  100%)",color:"#fff"},{"background-image":"linear-gradient(to right, rgb(59, 59, 59), rgb(34, 34, 34))",color:"#fff"},{"background-image":"linear-gradient(97deg, #7918c8, #f900ff)",color:"#fff"},{"background-image":"linear-gradient(220.55deg, #FFE70B 0%, #27B643 100%)",border:"1px solid rgb(182 214 19)",color:"#fff"}],c=l[Math.floor(Math.random()*l.length)];this._toastyService.showToast(this.toastMessages[o].title,this.toastMessages[o].message,{type:m.Custom,enableHtml:!0,customStyle:c})}randomToast(){let e=[this.basicToast,this.infoToast,this.successToast,this.WarningToast,this.ErrorToast,this.makeToast,this.htmlToast,this.htmlToast2,this.htmlToast3,this.customStyleToast,this.clickLongToast,this.clickLoadingToast,this.clickPromiseToast,this.clickEmojiToast,this.clickBeepToast,this.clickProgressBarToast];e[Math.floor(Math.random()*e.length)].bind(this)()}clickProgressBarToast(){this._toastyService.showToast("Progress","This toast has a progress bar",{type:m.Info,progressBar:!0})}clickBeepToast(){this._toastyService.showToast("Beep","This toast goes beep beep!",{type:m.Info,beep:!0})}clickLongToast(){this._toastyService.showToast("Long text","Lorem ipsum dolor sit amet, consectetur adipiscing elit. Nulla tincidunt euismod nunc quis suscipit. Curabitur sed semper tortor, at efficitur leo. Sed justo nulla, ullamcorper sed maximus a, vehicula ut nulla. Vivamus lobortis lacinia mi eu ultricies. Suspendisse quis sem ante. Vivamus consectetur lectus eget quam venenatis, sagittis aliquam lectus viverra. Donec nec molestie magna, ut ultricies tellus. Praesent faucibus urna imperdiet vulputate dapibus. Sed sit amet dolor odio. Donec est nulla, cursus sed aliquam at, cursus sit amet turpis. Nulla ut dolor eget dui laoreet hendrerit vel at tellus. Nullam aliquet sem at mauris scelerisque molestie. Nullam imperdiet blandit est et aliquam. Vestibulum euismod, libero id tincidunt rutrum, nulla augue aliquet risus, a rutrum elit orci eget lectus.")}clickLoadingToast(){this._toastyService.showToast("Wait","Process may take a few seconds to complete",{type:m.Custom,loading:!0})}clickEmojiToast(){this._toastyService.showToast("Error","\u2665\uFE0Flove emoji toast\u2665\uFE0F")}clickPromiseToast(){this._toastyService.showToastPromise("Fetching API",{promise:this.simulateApiCall(),loading:"Loading",success:e=>`Ok ${e.name}`,error:"Error"})}simulateApiCall(){return new Promise((e,n)=>setTimeout(()=>Math.random()<.5?e({name:"Toasty resolved"}):n(new Error("Toasty rejected")),2e3))}closeSidebar(){this.sidebarOpen=!1}static{this.\u0275fac=function(n){return new(n||t)(f(P))}}static{this.\u0275cmp=A({type:t,selectors:[["app-ng-toast-notify"]],standalone:!0,features:[O],decls:330,vars:61,consts:[[1,"flex","min-h-screen","bg-white","relative"],["id","overlay",1,"fixed","inset-0","bg-black","opacity-40","z-30","lg:hidden",3,"click"],["id","menu-btn",1,"fixed","lg:hidden","p-4","focus:outline-none","top-2","right-2","z-50",3,"click"],["xmlns","http://www.w3.org/2000/svg","fill","none","viewBox","0 0 24 24","stroke","currentColor",1,"h-6","w-6","text-gray-800"],["stroke-linecap","round","stroke-linejoin","round","stroke-width","2","d","M4 6h16M4 12h16m-7 6h7"],["id","sidebar",1,"fixed","top-0","left-0","w-64","p-6","space-y-6","h-full","bg-white","shadow-lg","overflow-y-auto","transform","transition-transform","duration-300","lg:translate-x-0","z-40"],[1,"text-xs","font-semibold","text-gray-400","uppercase","mb-2"],[1,"space-y-1"],["href","#install",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#addcomponent",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#useservice",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#basic-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#info-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#html-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#custom-css-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#long-message-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#loading-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#promise-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#emoji-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#swipe",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#beep",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#progressbar",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#sticky",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#container-settings",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],["href","#demo-toast",1,"block","px-2","py-1","rounded","hover:bg-gray-100",3,"click"],[1,"lg:ml-64","px-8","py-10","space-y-5","font-sans","text-sm"],[1,"text-center","space-y-4","w-5/6"],[1,"text-5xl","font-bold","tracking-tight"],[1,"text-lg","text-gray-600"],[1,"flex","gap-4","justify-center"],[1,"cursor-pointer","px-5","py-2","bg-emerald-500","hover:bg-emerald-600","text-white","rounded","shadow","font-medium",3,"click"],["href","https://github.com/rickmarina/ng-libraries","target","_blank",1,"px-5","py-2","border","border-gray-300","rounded","hover:bg-gray-100","transition","text-sm"],["id","install",1,"w-full","mx-auto"],[1,"text-2xl","font-semibold","mb-2"],[1,"block","text-sm","font-medium","text-gray-700","mb-2"],[3,"innerCode","run"],["id","addcomponent",1,"w-full","mx-auto"],[1,"block","text-sm","text-gray-700","mb-2"],["title","component.html",3,"innerCode","run"],[1,""],["href","#container-settings",1,"font-medium","text-blue-600","dark:text-blue-500","hover:underline"],[1,"w-full","mx-auto","my-1"],["title","component.ts",3,"innerCode","run"],["id","useservice",1,"w-full","mx-auto"],[1,"block","mb-2"],["title","Inject the service in your component",3,"innerCode","run"],["title","Use service to display toasts",3,"innerCode","run"],["id","basic-toast",1,"w-full","mx-auto"],["title","Basic toast",3,"innerCode","run"],["id","info-toast",1,"w-full","mx-auto"],[1,"b",3,"click"],[3,"innerCode"],["id","html-toast",1,"w-full","mx-auto"],[1,"text-sm","text-gray-600","mb-4"],[3,"title","innerCode"],["id","custom-css-toast",1,"w-full","max-auto"],["title","Custom red toast",3,"innerCode"],["title","Mango toast",3,"innerCode"],["title","Blackhole toast",3,"innerCode"],["id","long-message-toast",1,"w-full","max-auto"],["title","Long message toast",3,"innerCode"],["id","loading-toast",1,"w-full","max-auto"],["title","Loading",3,"innerCode"],["id","promise-toast",1,"w-full","max-auto"],[1,"w-4/6"],["title","Promise",3,"innerCode","run"],["id","emoji-toast",1,"w-full","max-auto"],["title","Emoji",3,"innerCode"],["id","swipe",1,"w-full","max-auto"],["title","Swipe to close",3,"innerCode","run"],["id","beep",1,"w-full","max-auto"],["title","Notify sound",3,"innerCode","run"],["id","progressbar",1,"w-full","max-auto"],["title","Progress bar",3,"innerCode","run"],["id","sticky",1,"w-full","max-auto"],["title","Sticky toast",3,"innerCode","run"],["id","container-settings",1,"w-full","max-auto"],["title","Position",3,"innerCode","run"],["title","Duration",3,"innerCode","run"],["title","Grouping",3,"innerCode","run"],["title","All settings together",3,"innerCode","run"],[1,"w-5/6"],[1,"font-medium"],[1,"overflow-x-auto"],[1,"min-w-full","text-sm","text-left","border","border-gray-200","rounded","overflow-hidden"],[1,"bg-gray-100","text-xs","uppercase","text-gray-600","border-b","border-gray-200"],[1,"px-4","py-2"],[1,"divide-y","divide-gray-200"],[1,"px-4","py-2","font-medium","text-gray-900"],[1,"px-4","py-2","text-gray-700"],[1,"px-4","py-2","text-gray-500"],["id","demo-toast",1,"w-full","space-y-4","mx-auto","hidden"],[1,"p-3","border","border-gray-200","rounded","space-y-3"],[1,"text-base","font-semibold"],[1,"flex","gap-2"],[1,"flex-1"],[1,"block","text-xs","font-medium","text-gray-700","mb-1"],["type","text","placeholder","Title",1,"w-full","px-3","py-1.5","border","border-gray-300","rounded",3,"ngModelChange","ngModel"],["rows","2","placeholder","Message",1,"w-full","px-3","py-1.5","border","border-gray-300","rounded","resize-none",3,"ngModelChange","ngModel"],[1,"flex","items-center","gap-3"],[1,"text-xs","font-medium","text-gray-700","whitespace-nowrap"],["type","range","min","1","max","10","step","1",1,"flex-1",3,"ngModelChange","ngModel"],[1,"flex","gap-3"],[1,"w-full","px-3","py-1.5","border","border-gray-300","rounded",3,"ngModelChange","ngModel"],["value","top-left"],["value","top"],["value","top-right"],["value","bottom-left"],["value","bottom"],["value","bottom-right"],[1,"flex-1","flex","items-center","gap-2"],[3,"position","duration","capacity","grouping"]],template:function(n,o){n&1&&(s(0,"div",0)(1,"div",1),d("click",function(){return o.sidebarOpen=!o.sidebarOpen}),r(),s(2,"button",2),d("click",function(){return o.sidebarOpen=!o.sidebarOpen}),xe(),s(3,"svg",3),p(4,"path",4),r()(),we(),s(5,"aside",5)(6,"div")(7,"h2",6),a(8,"Basics"),r(),s(9,"ul",7)(10,"li")(11,"a",8),d("click",function(){return o.closeSidebar()}),a(12,"Installation"),r()(),s(13,"li")(14,"a",9),d("click",function(){return o.closeSidebar()}),a(15,"Add component"),r()(),s(16,"li")(17,"a",10),d("click",function(){return o.closeSidebar()}),a(18,"Use service"),r()()()(),s(19,"div")(20,"h2",6),a(21,"Toasts types"),r(),s(22,"ul",7)(23,"li")(24,"a",11),d("click",function(){return o.closeSidebar()}),a(25,"Basic Toast"),r()(),s(26,"li")(27,"a",12),d("click",function(){return o.closeSidebar()}),a(28,"Info/Success/Warn/Error Toast"),r()(),s(29,"li")(30,"a",13),d("click",function(){return o.closeSidebar()}),a(31,"HTML Message"),r()(),s(32,"li")(33,"a",14),d("click",function(){return o.closeSidebar()}),a(34,"Custom CSS"),r()(),s(35,"li")(36,"a",15),d("click",function(){return o.closeSidebar()}),a(37,"Long Message"),r()(),s(38,"li")(39,"a",16),d("click",function(){return o.closeSidebar()}),a(40,"Loading Icon"),r()(),s(41,"li")(42,"a",17),d("click",function(){return o.closeSidebar()}),a(43,"Promise Toast"),r()(),s(44,"li")(45,"a",18),d("click",function(){return o.closeSidebar()}),a(46,"Emoji Toast"),r()()()(),s(47,"div")(48,"h2",6),a(49,"Toast config"),r(),s(50,"ul",7)(51,"li")(52,"a",19),d("click",function(){return o.closeSidebar()}),a(53,"Close on swipe"),r()(),s(54,"li")(55,"a",20),d("click",function(){return o.closeSidebar()}),a(56,"Beep sound"),r()(),s(57,"li")(58,"a",21),d("click",function(){return o.closeSidebar()}),a(59,"Progress bar"),r()(),s(60,"li")(61,"a",22),d("click",function(){return o.closeSidebar()}),a(62,"Sticky"),r()()()(),s(63,"div")(64,"h2",6),a(65,"Toasty Container"),r(),s(66,"ul",7)(67,"li")(68,"a",23),d("click",function(){return o.closeSidebar()}),a(69,"Settings"),r()()()(),s(70,"div")(71,"h2",6),a(72,"API"),r(),s(73,"ul",7)(74,"li")(75,"a",24),d("click",function(){return o.closeSidebar()}),a(76,"showToast()"),r()(),s(77,"li")(78,"a",24),d("click",function(){return o.closeSidebar()}),a(79,"showToastPromise()"),r()()()()(),s(80,"div",25)(81,"header",26)(82,"h1",27),a(83),r(),s(84,"p",28),a(85,"Lightweight and flexible toast notifications for Angular"),r(),s(86,"div",29)(87,"button",30),d("click",function(){return o.randomToast()}),a(88," Make a Toast "),r(),s(89,"a",31),a(90," GitHub \u2197 "),r()()(),s(91,"section",32)(92,"h2",33),a(93,"Installation"),r(),s(94,"label",34),a(95,"Install via npm"),r(),p(96,"app-code",35),r(),s(97,"section",36)(98,"h2",33),a(99,"Add component"),r(),s(100,"label",37),a(101,"Add this placeholder component with the following code in your html and check out you have the required imports in your component.ts"),r(),p(102,"app-code",38),s(103,"div",39),a(104,"The container can be customized for default settings: position, capacity and duration. See "),s(105,"a",40),a(106,"Container settings"),r(),a(107," to get detailed information."),r(),s(108,"div",41),p(109,"app-code",42),r()(),s(110,"section",43)(111,"h2",33),a(112,"Service"),r(),s(113,"label",44),a(114,"Add toasty service to your component and use it to display toasts."),r(),p(115,"app-code",45),s(116,"div")(117,"label",44),a(118,"Display a basic toast!"),r(),p(119,"app-code",46),r()(),s(120,"section",47)(121,"h2",33),a(122,"Basic Toast"),r(),s(123,"label",44),a(124,"This is the most elemental toast."),r(),p(125,"app-code",48),r(),s(126,"section",49)(127,"h2",33),a(128,"Info/Success/Warning/Error Toast"),r(),s(129,"label",44),a(130,"Some customized toasts with color and icon."),r(),s(131,"button",50),d("click",function(){return o.infoToast()}),a(132,"Info"),r(),s(133,"button",50),d("click",function(){return o.successToast()}),a(134,"Sucess"),r(),s(135,"button",50),d("click",function(){return o.WarningToast()}),a(136,"Warning"),r(),s(137,"button",50),d("click",function(){return o.ErrorToast()}),a(138,"Error"),r(),p(139,"app-code",51),r(),s(140,"section",52)(141,"h2",33),a(142,"HTML Message"),r(),s(143,"p",53),a(144,"You can use HTML content in your toast message if you indicate enableHtml: true as option."),r(),p(145,"app-code",54)(146,"app-code",54)(147,"app-code",54),r(),s(148,"section",55)(149,"h2",33),a(150,"Custom CSS Toast"),r(),a(151," It is possible to create custom css rules indicating type: Custom and customStyle parameters. "),p(152,"app-code",56)(153,"app-code",57)(154,"app-code",58),a(155," Note that you can define key:value pairs just as you would in a CSS document inside a variable and assign them to "),s(156,"b"),a(157,"customStyle: style"),r()(),s(158,"section",59)(159,"h2",33),a(160,"Long Message Toast"),r(),a(161," Toast is adapted vertically in base of the content length. "),p(162,"app-code",60),r(),s(163,"section",61)(164,"h2",33),a(165,"Loading Icon Toast"),r(),a(166," Aditionally you can add a loding icon, just setting loading: true in the configuration parameter. "),p(167,"app-code",62),r(),s(168,"section",63)(169,"h2",33),a(170,"Promise Toast"),r(),s(171,"div",64),a(172,"It is possible to associate a promise. For this task you should use showToastPromise method that provides two arguments; a title and a ToastyPromise model. You can see in the following code an example."),r(),p(173,"app-code",65),s(174,"button",50),d("click",function(){return o.clickPromiseToast()}),a(175,"Promise toast example"),r()(),s(176,"section",66)(177,"h2",33),a(178,"Emoji Toast"),r(),s(179,"div",64),a(180,"Add emoji in the toasty simple using it directly in the message body:"),r(),p(181,"app-code",67),r(),s(182,"section",68)(183,"h2",33),a(184,"Close on swipe"),r(),s(185,"div",64),a(186,"You can close any toast by swiping it horizontally. This is useful for mobile devices."),r(),p(187,"app-code",69),r(),s(188,"section",70)(189,"h2",33),a(190,"Beep sound"),r(),s(191,"div",64),a(192,"You can enable a beep sound when a toast is displayed. This is useful for notifications that require attention. By default beeps are false"),r(),p(193,"app-code",71),r(),s(194,"section",72)(195,"h2",33),a(196,"Progress Bar"),r(),s(197,"div",64),a(198,"You can add a progress bar to your toast. It will animate while the toast is visible."),r(),p(199,"app-code",73),r(),s(200,"section",74)(201,"h2",33),a(202,"Sticky Toast"),r(),s(203,"div",64),a(204,"You can create a sticky toast that will not disappear automatically. The user must close it manually."),r(),p(205,"app-code",75),r(),s(206,"section",76)(207,"h2",33),a(208,"Container settings"),r(),s(209,"div",64),a(210,"You can set some default settinngs for toasts container."),r(),p(211,"app-code",77)(212,"app-code",78)(213,"app-code",78)(214,"app-code",79)(215,"app-code",80),s(216,"section",81)(217,"span",82),a(218,"Component parameters"),r(),s(219,"div",83)(220,"table",84)(221,"thead",85)(222,"tr")(223,"th",86),a(224,"Name"),r(),s(225,"th",86),a(226,"Type"),r(),s(227,"th",86),a(228,"Default"),r(),s(229,"th",86),a(230,"Description"),r()()(),s(231,"tbody",87)(232,"tr")(233,"td",88),a(234,"position"),r(),s(235,"td",89),a(236,"string"),r(),s(237,"td",90),a(238,'"bottom-right"'),r(),s(239,"td",89),a(240,"Controls the toast position. Options: "),s(241,"code"),a(242,"top-left"),r(),a(243,", "),s(244,"code"),a(245,"top"),r(),a(246,", "),s(247,"code"),a(248,"top-right"),r(),a(249,", "),s(250,"code"),a(251,"bottom-left"),r(),a(252,", "),s(253,"code"),a(254,"bottom"),r(),a(255,", "),s(256,"code"),a(257,"bottom-right"),r(),a(258,". "),r()(),s(259,"tr")(260,"td",88),a(261,"duration"),r(),s(262,"td",89),a(263,"number"),r(),s(264,"td",90),a(265,"5"),r(),s(266,"td",89),a(267,"Time in seconds before a toast disappears automatically."),r()(),s(268,"tr")(269,"td",88),a(270,"capacity"),r(),s(271,"td",89),a(272,"number"),r(),s(273,"td",90),a(274,"10"),r(),s(275,"td",89),a(276,"Maximum number of concurrent toasts allowed."),r()(),s(277,"tr")(278,"td",88),a(279,"grouping"),r(),s(280,"td",89),a(281,"boolean"),r(),s(282,"td",90),a(283,"false"),r(),s(284,"td",89),a(285,"Allow same toasts group together"),r()()()()()()(),s(286,"section",91)(287,"h2",33),a(288,"Demo"),r(),s(289,"div",92)(290,"h3",93),a(291,"Toasty settings"),r(),s(292,"div",94)(293,"div",95)(294,"label",96),a(295,"Title"),r(),s(296,"input",97),k("ngModelChange",function(c){return I(o.title,c)||(o.title=c),c}),r()(),s(297,"div",95)(298,"label",96),a(299,"Message"),r(),s(300,"textarea",98),k("ngModelChange",function(c){return I(o.message,c)||(o.message=c),c}),r()()(),s(301,"div",99)(302,"label",100),a(303),r(),s(304,"input",101),k("ngModelChange",function(c){return I(o.durationToast,c)||(o.durationToast=c),c}),r()()(),s(305,"div",92)(306,"h3",93),a(307,"Toasty container settings"),r(),s(308,"div",102)(309,"div",95)(310,"label",96),a(311,"Position"),r(),s(312,"select",103),k("ngModelChange",function(c){return I(o.containerPosition,c)||(o.containerPosition=c),c}),s(313,"option",104),a(314,"Top left"),r(),s(315,"option",105),a(316,"Top"),r(),s(317,"option",106),a(318,"Top right"),r(),s(319,"option",107),a(320,"Bottom left"),r(),s(321,"option",108),a(322,"Bottom"),r(),s(323,"option",109),a(324,"Bottom right"),r()()(),s(325,"div",110)(326,"label",100),a(327),r(),s(328,"input",101),k("ngModelChange",function(c){return I(o.containerCapacity,c)||(o.containerCapacity=c),c}),r()()()()(),p(329,"toasty",111),r()()),n&2&&(u(),x("hidden",!o.sidebarOpen),u(4),x("-translate-x-full",!o.sidebarOpen),u(78),S(o.libName),u(13),h("innerCode","npm install "+o.libName)("run",!1),u(6),h("innerCode","<toasty></toasty>")("run",!1),u(7),h("innerCode",o.componentcode)("run",!1),u(6),h("innerCode","constructor(private _toastyService: ToastyService) { }")("run",!1),u(4),h("innerCode","this._toastyService.showToast('Title', 'Message body');")("run",!0),u(6),h("innerCode","this._toastyService.showToast('Title', 'Message body');")("run",!0),u(14),h("innerCode",o.infosuccessetctoasts),u(6),h("title","Basic toast with html list")("innerCode","this._toastyService.showToast('Html', '<b>Details list</b><ul><li>item 1</li><li>item 2</li></ul>', { type: ToastType.Basic, enableHtml: true });"),u(),h("title","Success toast with html content & gif")("innerCode","this._toastyService.showToast('Html', 'Data successfully saved. <img width=80 src=./assets/img/dancing-spiderman.gif />', { type: ToastType.Success, enableHtml: true });"),u(),h("title","Info toast example with achievement unlocked")("innerCode","this._toastyService.showToast('Wizard', 'Achievement unlocked. <img width=100 src=./assets/img/wizard.gif />', { type: ToastType.Info, enableHtml: true });"),u(5),h("innerCode",o.customcode),u(),h("innerCode",o.customcode2),u(),h("innerCode",o.customcode3),u(8),h("innerCode",o.longToast),u(5),h("innerCode",o.loadingToast),u(6),h("innerCode",o.promiseToast)("run",!1),u(8),h("innerCode","this._toastyService.showToast('Error', '\u2665\uFE0Flove emoji toast\u2665\uFE0F');"),u(6),h("innerCode","this._toastyService.showToast('Swipe', 'Swipe me to close');")("run",!0),u(6),h("innerCode","this._toastyService.showToast('Beep', 'This toast will beep',{ beep: true });")("run",!0),u(6),h("innerCode","this._toastyService.showToast('Progress', 'This toast has a progress bar', { progressBar: true });")("run",!0),u(6),h("innerCode","this._toastyService.showToast('Sticky', 'This toast is sticky and will not disappear automatically.', { sticky: true });")("run",!0),u(6),h("innerCode",'<toasty [position]="containerPosition"></toasty>')("run",!1),u(),h("innerCode",'<toasty [duration]="containerDuration"></toasty>')("run",!1),u(),h("innerCode",'<toasty [capacity]="containerCapacity"></toasty>')("run",!1),u(),h("innerCode",'<toasty [capacity]="containerGrouping"></toasty>')("run",!1),u(),h("innerCode",'<toasty [position]="containerPosition" [duration]="containerDuration" [capacity]="containerCapacity" [grouping]="containerGrouping"></toasty>')("run",!1),u(81),F("ngModel",o.title),u(4),F("ngModel",o.message),u(3),q("Duration: ",o.durationToast,"s"),u(),F("ngModel",o.durationToast),u(8),F("ngModel",o.containerPosition),u(15),q("Capacity: ",o.containerCapacity,""),u(),F("ngModel",o.containerCapacity),u(),h("position",o.containerPosition)("duration",o.containerDuration)("capacity",o.containerCapacity)("grouping",o.containerGrouping))},dependencies:[Ue,pt,dt,mt,re,be,ae,at,Ce,We]})}}return t})();export{Ui as NgToastNotifyComponent};
