+++ docs/main-KPGSNRMY.js
@@ -1,0 +1,1 @@
+import{a as r,b as i,d as m,e as a}from"./chunk-NGEOEQX2.js";import{Ta as n,Y as e,fb as p}from"./chunk-RBV3YYQO.js";var s=[{path:"",loadComponent:()=>import("./chunk-6DXUKTKW.js").then(o=>o.HomeComponent)},{path:"ng-toast-notify",loadComponent:()=>import("./chunk-GKGHKTOS.js").then(o=>o.NgToastNotifyComponent)}];var f={providers:[m(s,a())]};var c=(()=>{class o{static{this.\u0275fac=function(t){return new(t||o)}}static{this.\u0275cmp=e({type:o,selectors:[["app-root"]],standalone:!0,features:[p],decls:1,vars:0,template:function(t,l){t&1&&n(0,"router-outlet")},dependencies:[i]})}}return o})();r(c,f).catch(o=>console.error(o));
