+++ projects/ng-toast-notify/package.json
@@ -1,6 +1,6 @@
 {
   "name": "ng-toast-notify",
-  "version": "1.0.7",
+  "version": "1.0.8",
   "description": "Lightweight and flexible toast notifications for Angular ",
   "keywords": [
     "angular",
